% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{bcftools_parse_single_vcf}
\alias{bcftools_parse_single_vcf}
\title{Parse Single-Sample VCF with bcftools}
\usage{
bcftools_parse_single_vcf(vcf)
}
\arguments{
\item{vcf}{VCF with a single sample.}
}
\value{
Tibble with all the main, FORMAT, and INFO fields detected in the VCF header as columns.
}
\description{
Parse Single-Sample VCF with bcftools
}
