% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{bcftools_parse_vcf_regions}
\alias{bcftools_parse_vcf_regions}
\title{Parse VCF regions with bcftools}
\usage{
bcftools_parse_vcf_regions(vcf, r)
}
\arguments{
\item{vcf}{Path to VCF. Can be S3, http or local. If presigned URL, need to
also concatenate the VCF index as in 'vcf_url##idx##vcfi_url'.}

\item{r}{Character vector of regions to subset (e.g. c('chr1:123-456', 'chr2:789-1000'))}
}
\value{
A tibble with all the main, FORMAT, and INFO fields detected in
the VCF header as columns, for the regions specified in \code{r} (if any).
}
\description{
Parses VCF regions with bcftools. The VCF subset is written to a temporary
file in the local filesystem, then parsed into a tibble object.
}
\examples{
\dontrun{
vcf_local <- here::here("nogit/tso/2023-05-30/SBJ00595_L2100178/PTC_SrSqCMM1pc_L2100178_rerun_MergedSmallVariants.vcf.gz")
r <- c("chr1:115256529-115256529", "chr2:29443613-29443613")
bcftools_parse_vcf_regions(vcf_local, r)
}
}
