% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcftools_stats.R
\name{BcftoolsStatsFile}
\alias{BcftoolsStatsFile}
\title{BcftoolsStatsFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{bcftools_stats.txt} file (QUAL section) output from running \verb{bcftools stats}.
}
\examples{
\dontrun{
x <- "/path/to/bcftools_stats.txt"
d <- BcftoolsStatsFile$new(x)
d_parsed <- d$read()
d$write(d_parsed, out_dir = tempdir(), prefix = "sample123", out_format = "tsv")
}
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{BcftoolsStatsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BcftoolsStatsFile-read}{\code{BcftoolsStatsFile$read()}}
\item \href{#method-BcftoolsStatsFile-write}{\code{BcftoolsStatsFile$write()}}
\item \href{#method-BcftoolsStatsFile-clone}{\code{BcftoolsStatsFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BcftoolsStatsFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-BcftoolsStatsFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the QUAL section from the \code{bcftools_stats.txt} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BcftoolsStatsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BcftoolsStatsFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-BcftoolsStatsFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{bcftools_stats.txt} (only QUAL section)
file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BcftoolsStatsFile$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BcftoolsStatsFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BcftoolsStatsFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BcftoolsStatsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
