% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclconvert.R
\name{bclconvert_read_adaptermetrics}
\alias{bclconvert_read_adaptermetrics}
\title{BCLConvert Adapter Metrics}
\usage{
bclconvert_read_adaptermetrics(x)
}
\arguments{
\item{x}{Path to \code{Adapter_Metrics.csv} file.}
}
\value{
Tibble
}
\description{
Reads the \code{Adapter_Metrics.csv} file in the \code{Reports} directory
output by BCLConvert.
}
\examples{
\dontrun{
x <- here::here("nogit/bcl_convert/WGS_TsqNano/Reports/Adapter_Metrics.csv")
bclconvert_read_adaptermetrics(x)
}
}
