% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclconvert.R
\name{bclconvert_read_demultiplexstats}
\alias{bclconvert_read_demultiplexstats}
\title{BCLConvert Demultiplex Stats}
\usage{
bclconvert_read_demultiplexstats(x)
}
\arguments{
\item{x}{Path to \code{Demultiplex_Stats.csv} file.}
}
\value{
Tibble
}
\description{
Reads the \code{Demultiplex_Stats.csv} file in the \code{Reports} directory
output by BCLConvert.
}
\examples{
\dontrun{
x <- here::here("nogit/bcl_convert/WGS_TsqNano/Reports/Demultiplex_Stats.csv")
bclconvert_read_demultiplexstats(x)
}
}
