% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclconvert.R
\name{bclconvert_read_indexhoppingcounts}
\alias{bclconvert_read_indexhoppingcounts}
\title{BCLConvert Index Hopping Counts}
\usage{
bclconvert_read_indexhoppingcounts(x)
}
\arguments{
\item{x}{Path to \code{Index_Hopping_Counts.csv} file.}
}
\value{
Tibble
}
\description{
Reads the \code{Index_Hopping_Counts.csv} file in the \code{Reports} directory
output by BCLConvert.
}
\examples{
\dontrun{
x <- here::here("nogit/bcl_convert/WGS_TsqNano/Reports/Index_Hopping_Counts.csv")
bclconvert_read_indexhoppingcounts(x)
}
}
