% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclconvert.R
\name{bclconvert_read_topunknownbarcodes}
\alias{bclconvert_read_topunknownbarcodes}
\title{BCLConvert Top Unknown Barcodes}
\usage{
bclconvert_read_topunknownbarcodes(x)
}
\arguments{
\item{x}{Path to \code{Top_Unknown_Barcodes.csv} file.}
}
\value{
Tibble
}
\description{
Reads the \code{Top_Unknown_Barcodes.csv} file in the \code{Reports} directory
output by BCLConvert.
}
\examples{
\dontrun{
x <- here::here("nogit/bcl_convert/WGS_TsqNano/Reports/Top_Unknown_Barcodes.csv")
bclconvert_read_topunknownbarcodes(x)
}
}
