% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclconvert.R
\name{BclconvertReports}
\alias{BclconvertReports}
\title{BclconvertReports R6 Class}
\description{
Reads and writes tidy versions of files within the \code{Reports} directory output
from BCLConvert.
}
\examples{
\dontrun{
b <- BclconvertReports$new(here::here("nogit/bcl_convert/WGS_TsqNano/Reports"))
b$path
b$contents
d <- b$read()
b$write(d, out_dir = tempdir(), prefix = "sampleA", out_format = "tsv")
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Path to the \code{Reports} directory.}

\item{\code{contents}}{Tibble with file path, basename, and size.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BclconvertReports-new}{\code{BclconvertReports$new()}}
\item \href{#method-BclconvertReports-print}{\code{BclconvertReports$print()}}
\item \href{#method-BclconvertReports-read}{\code{BclconvertReports$read()}}
\item \href{#method-BclconvertReports-write}{\code{BclconvertReports$write()}}
\item \href{#method-BclconvertReports-clone}{\code{BclconvertReports$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports-new"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports-new}{}}}
\subsection{Method \code{new()}}{
Create a new BclconvertReports object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports$new(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the \code{Reports} directory.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports-print"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports-print}{}}}
\subsection{Method \code{print()}}{
Print details about the BclconvertReports directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports-read"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports-read}{}}}
\subsection{Method \code{read()}}{
Reads contents of \code{Reports} directory output by BCLConvert.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of tibbles.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports-write"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports-write}{}}}
\subsection{Method \code{write()}}{
Writes tidied contents of \code{Reports} directory output by BCLConvert.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
