% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclconvert.R
\name{BclconvertReports375}
\alias{BclconvertReports375}
\title{BclconvertReports375 R6 Class}
\description{
Reads and writes tidy versions of files within the \code{Reports} directory output
from BCLConvert v4.2.7. See the DRAGEN v4.2 documentation at
https://support-docs.illumina.com/SW/dragen_v42/Content/SW/DRAGEN/OutputFiles.htm.
}
\examples{
\dontrun{
p1 <- "nogit/bcl_convert/WGS_TsqNano/Reports"
b <- here::here(p1) |>
  BclconvertReports375$new()
b$path
b$contents
d <- b$read()
b$write(d, out_dir = tempdir(), prefix = "sampleA", out_format = "tsv")
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Path to the \code{Reports} directory.}

\item{\code{contents}}{Tibble with file path, basename, and size.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BclconvertReports375-new}{\code{BclconvertReports375$new()}}
\item \href{#method-BclconvertReports375-print}{\code{BclconvertReports375$print()}}
\item \href{#method-BclconvertReports375-read_adaptermetrics}{\code{BclconvertReports375$read_adaptermetrics()}}
\item \href{#method-BclconvertReports375-read_demultiplexstats}{\code{BclconvertReports375$read_demultiplexstats()}}
\item \href{#method-BclconvertReports375-read_indexhoppingcounts}{\code{BclconvertReports375$read_indexhoppingcounts()}}
\item \href{#method-BclconvertReports375-read_topunknownbarcodes}{\code{BclconvertReports375$read_topunknownbarcodes()}}
\item \href{#method-BclconvertReports375-read_fastqlist}{\code{BclconvertReports375$read_fastqlist()}}
\item \href{#method-BclconvertReports375-read}{\code{BclconvertReports375$read()}}
\item \href{#method-BclconvertReports375-write}{\code{BclconvertReports375$write()}}
\item \href{#method-BclconvertReports375-clone}{\code{BclconvertReports375$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports375-new"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports375-new}{}}}
\subsection{Method \code{new()}}{
Create a new BclconvertReports375 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports375$new(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the \code{Reports} directory.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports375-print"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports375-print}{}}}
\subsection{Method \code{print()}}{
Print details about the BclconvertReports375 directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports375$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports375-read_adaptermetrics"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports375-read_adaptermetrics}{}}}
\subsection{Method \code{read_adaptermetrics()}}{
Read Adapter_Metrics.csv file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports375$read_adaptermetrics(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to Adapter_Metrics.csv file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports375-read_demultiplexstats"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports375-read_demultiplexstats}{}}}
\subsection{Method \code{read_demultiplexstats()}}{
Read Demultiplex_Stats.csv file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports375$read_demultiplexstats(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to Demultiplex_Stats.csv file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports375-read_indexhoppingcounts"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports375-read_indexhoppingcounts}{}}}
\subsection{Method \code{read_indexhoppingcounts()}}{
Read Index_Hopping_Counts.csv file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports375$read_indexhoppingcounts(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to Index_Hopping_Counts.csv file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports375-read_topunknownbarcodes"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports375-read_topunknownbarcodes}{}}}
\subsection{Method \code{read_topunknownbarcodes()}}{
Read Top_Unknown_Barcodes.csv file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports375$read_topunknownbarcodes(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to Top_Unknown_Barcodes.csv file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports375-read_fastqlist"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports375-read_fastqlist}{}}}
\subsection{Method \code{read_fastqlist()}}{
Read fastq_list.csv file.
\itemize{
\item rgid: read group.
\item rgsm: sample ID.
\item rglb: library.
\item lane: flow cell lane.
\item readnum: read number (1 or 2).
\item filepath: path to the FASTQ file.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports375$read_fastqlist(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to fastq_list.csv file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports375-read"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports375-read}{}}}
\subsection{Method \code{read()}}{
Reads contents of \code{Reports} directory output by BCLConvert.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports375$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of tibbles.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports375-write"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports375-write}{}}}
\subsection{Method \code{write()}}{
Writes tidied contents of \code{Reports} directory output by BCLConvert.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports375$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BclconvertReports375-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BclconvertReports375-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BclconvertReports375$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
