% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverageMetricsFile.R
\name{CoverageMetricsFile}
\alias{CoverageMetricsFile}
\title{CoverageMetricsFile R6 Class}
\description{
Contains methods for reading and displaying contents of
the \verb{wgs_coverage_metrics_<phenotype>.csv} file output from DRAGEN.
This file contains read depth of coverage metrics.
}
\examples{
x1 <- system.file("extdata/SEQC-II.wgs_coverage_metrics_normal.csv", package = "dracarys")
x2 <- system.file("extdata/SEQC-II.wgs_coverage_metrics_tumor.csv", package = "dracarys")
cm1 <- CoverageMetricsFile$new(x1)
cm2 <- CoverageMetricsFile$new(x2)
read(cm1)
read(cm2)

}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{CoverageMetricsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-read}{\code{CoverageMetricsFile$read()}}
\item \href{#method-clone}{\code{CoverageMetricsFile$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname">}\href{../../dracarys/html/File.html#method-bname}{\code{dracarys::File$bname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize">}\href{../../dracarys/html/File.html#method-initialize}{\code{dracarys::File$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print">}\href{../../dracarys/html/File.html#method-print}{\code{dracarys::File$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type">}\href{../../dracarys/html/File.html#method-type}{\code{dracarys::File$type()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Reads the \verb{wgs_coverage_metrics_<phenotype>.csv} file output from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverageMetricsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item label: file label.
\item var: variable name.
\item var_abbrev: variable abbreviation.
\item pct: percentage value.
\item count: count value.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverageMetricsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
