% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal_meta.R
\name{cttso_metadata}
\alias{cttso_metadata}
\title{Metadata for cttso workflow}
\usage{
cttso_metadata(pmeta, status = c("Succeeded"))
}
\arguments{
\item{pmeta}{Portal workflows metadata table in TSV format.}

\item{status}{Workflow status to keep (default: Succeeded).}
}
\value{
A tibble with metadata per cttso workflow run.
}
\description{
Metadata for cttso workflow
}
\examples{
\dontrun{
pmeta <- here::here("nogit/data_portal/2023-05-07_workflows_5f43da12-0b6a-41ce-86c9-9ee62df8792e.csv")
cttso_metadata(pmeta)
}
}
