% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_test}
\alias{db_test}
\title{Test DB}
\usage{
db_test(path, prefix, portalRunId)
}
\description{
Test DB
}
\examples{
\dontrun{
path <- file.path(
  "~/s3/pipeline-prod-cache-503977275616-ap-southeast-2/byob-icav2/production",
  "analysis/wgts-qc/20241123ffa837c4/L2401621_dragen_alignment"
)
prefix <-  "L2401621"
outdir <- path
max_files <- 1000
portalRunId <- "abcd1234"
}

}
