% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{dr_func_eval}
\alias{dr_func_eval}
\title{Evaluate dracarys Function}
\usage{
dr_func_eval(f, v = NULL)
}
\arguments{
\item{f}{Name of function to evaluate.}

\item{v}{Character vector of strings evaluating to functions. By default,
this points to the functions in the DR_FILE_REGEX dracarys tibble.}
}
\value{
Evaluated function.
}
\description{
This is somewhat a hack for getting a function to evaluate based on a lookup
vector. If the function is not found, it returns NULL.
}
\examples{
mean_1_to_10 <- dr_func_eval("mean", v = c("mean", "sd"))(1:10)
x <- system.file("extdata/tso/sample705.fragment_length_hist.json.gz", package = "dracarys")
obj <- dr_func_eval("TsoFragmentLengthHistFile")$new(x)
}
