% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{dr_gds_download}
\alias{dr_gds_download}
\title{dracarys GDS Download}
\usage{
dr_gds_download(
  gdsdir,
  outdir,
  token = Sys.getenv("ICA_ACCESS_TOKEN"),
  pattern = NULL,
  dryrun = FALSE
)
}
\arguments{
\item{gdsdir}{Full path to GDS directory, must end with \code{/}.}

\item{outdir}{Path to output directory.}

\item{token}{ICA access token (by default uses $ICA_ACCESS_TOKEN env var).}

\item{pattern}{Pattern to further filter the returned file type tibble.}

\item{dryrun}{Just list the files that will be downloaded?}
}
\description{
Download only GDS files that can be processed by dracarys.
}
