% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_s3.R
\name{dr_s3_download}
\alias{dr_s3_download}
\title{dracarys S3 Download}
\usage{
dr_s3_download(
  s3dir,
  outdir,
  max_objects = 100,
  pattern = NULL,
  regexes = DR_FILE_REGEX,
  dryrun = FALSE
)
}
\arguments{
\item{s3dir}{S3 directory.}

\item{outdir}{Path to output directory.}

\item{max_objects}{Maximum objects returned. By default, iterates by 1000
objects at a time until the bucket/prefix is exhausted.}

\item{pattern}{Pattern to further filter the returned file type tibble.}

\item{regexes}{Tibble with \code{regex} and \code{fun}ction name.}

\item{dryrun}{If TRUE, just list the files that will be downloaded (don't
download them).}
}
\description{
Download only S3 files that can be processed by dracarys.
}
\examples{
\dontrun{
p1 <- "s3://org.umccr.data.oncoanalyser/analysis_data/SBJ05373/sash"
p2 <- "20240707becde493/L2401018_L2401017/SBJ05373_MDX240220"
s3dir <- file.path(p1, p2)
regexes <- tibble::tribble(
  ~regex, ~fun,
  "multiqc_data\\\\.json$", "MultiqcJsonFile",
  "pcgr.*\\\\.json\\\\.gz$", "pcgrjson"
)
outdir <- sub("s3:/", "~/s3", s3dir)
dr_s3_download(s3dir = s3dir, outdir = outdir, max_objects = 500, regexes = regexes, dryrun = F)
}
}
