% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dracarys-package.R
\docType{package}
\name{dracarys-package}
\alias{dracarys}
\alias{dracarys-package}
\title{dracarys: Post-processing DRAGEN workflows}
\description{
Post-processing DRAGEN workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/umccr/dracarys}
  \item \url{https://umccr.github.io/dracarys/}
  \item Report bugs at \url{https://github.com/umccr/dracarys/issues}
}

}
\author{
\strong{Maintainer}: Peter Diakumis \email{peterdiakumis@gmail.com} (\href{https://orcid.org/0000-0002-7502-7545}{ORCID})

Authors:
\itemize{
  \item Victor San Kho Lin (\href{https://orcid.org/0000-0002-3940-4729}{ORCID}) [contributor]
}

}
\keyword{internal}
