% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{dracarys_multiqc}
\alias{dracarys_multiqc}
\title{Dracarys MultiQC Tidy}
\usage{
dracarys_multiqc(json, prefix, outdir, out_format = "tsv")
}
\arguments{
\item{json}{Path to \code{multiqc_data.json}.}

\item{prefix}{Prefix for output files.}

\item{outdir}{Path to output results.}

\item{out_format}{Format of output (tsv, parquet, both) (def: tsv).}
}
\value{
Generates TSV and/or Parquet representations of the input
MultiQC JSON file.
}
\description{
Generate tidier representations of MultiQC JSON output
}
