% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{dracarys_tso}
\alias{dracarys_tso}
\title{Dracarys Tso Tidy}
\usage{
dracarys_tso(indir, outdir, out_format = "tsv")
}
\arguments{
\item{indir}{Path to TSO500 ctDNA workflow results.}

\item{outdir}{Path to output dracarys results.}

\item{out_format}{Format of output (tsv, parquet, both) (def: tsv).}
}
\value{
Generates tidy TSV and/or Parquet representations of the workflow results.
}
\description{
Generate tidier representations of TSO500 ctDNA workflow outputs.
}
