% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_cnv_metrics_read}
\alias{dragen_cnv_metrics_read}
\title{Read DRAGEN CNV Metrics}
\usage{
dragen_cnv_metrics_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble with metrics.
}
\description{
Reads the \code{cnv_metrics.csv} file output from DRAGEN.
}
\examples{
\dontrun{
dragen_cnv_metrics_read(x)
}

}
