% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_contig_mean_coverage_plot}
\alias{dragen_contig_mean_coverage_plot}
\title{Plot DRAGEN Contig Mean Coverage
TODO
Plots the \verb{wgs_contig_mean_cov_<phenotype>.csv} files.}
\usage{
dragen_contig_mean_coverage_plot(d, top_alt_n = 15)
}
\arguments{
\item{d}{Parsed tibble.}

\item{top_alt_n}{Number of top covered alt contigs to plot per phenotype.}
}
\value{
A ggplot2 object with chromosomes on X axis, and coverage on Y axis.
}
\description{
Plot DRAGEN Contig Mean Coverage
TODO
Plots the \verb{wgs_contig_mean_cov_<phenotype>.csv} files.
}
