% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_coverage_metrics_read}
\alias{dragen_coverage_metrics_read}
\title{Read DRAGEN Coverage Metrics}
\usage{
dragen_coverage_metrics_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble with metrics.
}
\description{
Reads the \code{coverage_metrics.csv} file generated by DRAGEN.
}
