% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_covreportbed_read}
\alias{dragen_covreportbed_read}
\title{Read DRAGEN \code{read_cov_report.bed} File}
\usage{
dragen_covreportbed_read(x)
}
\arguments{
\item{x}{Path to file}
}
\description{
Reads coveraage report file.
}
