% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen_fastqc.R
\name{dragen_fastqc_metrics_read}
\alias{dragen_fastqc_metrics_read}
\title{DRAGEN FASTQC Metrics}
\usage{
dragen_fastqc_metrics_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble with metrics.
}
\description{
Read DRAGEN \code{fastqc_metrics.csv} file.
}
