% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_gc_metrics_read}
\alias{dragen_gc_metrics_read}
\title{Read DRAGEN GC Metrics}
\usage{
dragen_gc_metrics_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble with metrics.
}
\description{
Reads the \code{gc_metrics.csv} file output from DRAGEN.
}
