% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_hethomratio_read}
\alias{dragen_hethomratio_read}
\title{Read DRAGEN Heterozygous/Homozygous Ratio File}
\usage{
dragen_hethomratio_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble with metrics.
}
\description{
Reads the \code{vc_hethom_ratio_metrics.csv} file output from DRAGEN.
}
