% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_msi_read}
\alias{dragen_msi_read}
\title{Read DRAGEN `microsat_output.json`` File}
\usage{
dragen_msi_read(x)
}
\arguments{
\item{x}{Path to file}
}
\description{
Reads \code{msi.json.gz} file output from the TSO500 workflow.
}
