% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_ploidy_estimation_metrics_read}
\alias{dragen_ploidy_estimation_metrics_read}
\title{Read DRAGEN Ploidy Estimation Metrics}
\usage{
dragen_ploidy_estimation_metrics_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble with metrics.
}
\description{
Reads the \code{ploidy_estimation_metrics.csv} file generated by DRAGEN.
}
