% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_ploidy_vcf_read}
\alias{dragen_ploidy_vcf_read}
\title{Read DRAGEN Ploidy VCF}
\usage{
dragen_ploidy_vcf_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble with depth of coverage (dc) and normalised depth of coverage (ndc)
for each chromosome.
}
\description{
Reads the \code{ploidy.vcf.gz} file generated by DRAGEN.
}
