% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_rohbed_read}
\alias{dragen_rohbed_read}
\title{Read DRAGEN \code{roh.bed} File}
\usage{
dragen_rohbed_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble with roh regions, score, and number of homs/hets.
}
\description{
Reads roh bed file.
}
