% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_rohmetrics_read}
\alias{dragen_rohmetrics_read}
\title{Read DRAGEN \code{roh_metrics.csv} File}
\usage{
dragen_rohmetrics_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble with pct_snvs, n_regions.
}
\description{
Reads roh metrics file.
}
