% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_subprefix}
\alias{dragen_subprefix}
\title{DRAGEN File Subprefix}
\usage{
dragen_subprefix(x, suffix)
}
\arguments{
\item{x}{File name.}

\item{suffix}{Suffix to remove.}
}
\value{
Clean string.
}
\description{
Extracts a file subprefix for better table naming.
}
\examples{
x1 <- "L2401290.exon_contig_mean_cov.csv"
x2 <- "L2401290.tmb_contig_mean_cov.csv.gz"
x3 <- "foo.bar.exon_contig_mean_cov.csv.gz"
(s1 <- dragen_subprefix(x1, "_contig_mean_cov"))
(s2 <- dragen_subprefix(x2, "_contig_mean_cov"))
(s3 <- dragen_subprefix(x3, "_contig_mean_cov"))
}
