% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_sv_metrics_read}
\alias{dragen_sv_metrics_read}
\title{Read DRAGEN SV Metrics}
\usage{
dragen_sv_metrics_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble with metrics.
}
\description{
Reads the \code{sv_metrics.csv} file output from DRAGEN.
}
\examples{
\dontrun{
dragen_sv_metrics_read(x)
}
}
