% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_targeted_vcf_read}
\alias{dragen_targeted_vcf_read}
\title{Read DRAGEN Targeted VCF}
\usage{
dragen_targeted_vcf_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
Tibble.
}
\description{
Reads the \code{targeted.vcf.gz} file generated by DRAGEN.
}
