% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dragen_tmb_metrics_read}
\alias{dragen_tmb_metrics_read}
\title{Read DRAGEN \code{tmb.metrics.csv} File}
\usage{
dragen_tmb_metrics_read(x)
}
\arguments{
\item{x}{Path to file}
}
\description{
Reads tmb metrics file.
}
