% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsov2.R
\name{dtw_Wf_cttsov2}
\alias{dtw_Wf_cttsov2}
\title{Wf_cttsov2 Download Tidy and Write}
\usage{
dtw_Wf_cttsov2(
  path,
  prefix,
  outdir,
  outdir_tidy = file.path(outdir, "dracarys_tidy"),
  format = "rds",
  max_files = 1000,
  dryrun = FALSE,
  drid = NULL,
  dbconn = NULL
)
}
\arguments{
\item{path}{Path to directory with raw workflow results (S3 or
local filesystem).}

\item{prefix}{The LibraryID prefix of the sample (needed for path lookup).}

\item{outdir}{Path to output directory with raw files.}

\item{outdir_tidy}{Path to output directory with tidy files.}

\item{format}{Format of output files.}

\item{max_files}{Max number of files to list.}

\item{dryrun}{If TRUE, just list the files that will be downloaded (don't
download them).}

\item{drid}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}

\item{dbconn}{Database connection object.}
}
\value{
Tibble of tidy tibbles.
}
\description{
Downloads files from the \code{cttsov2} workflow and writes them in a tidy format.
}
\examples{
\dontrun{
#---- Local ----#
path <- file.path(
  "~/s3/pipeline-prod-cache-503977275616-ap-southeast-2/byob-icav2/production",
  "analysis/cttsov2/20250308f448d4e0"
)
prefix <- "L2500183"
outdir <- file.path(path, "dracarys_tidy")
drid <- "abcde123"
drid <- "fghij456"
dbconn <- DBI::dbConnect(
  drv = RPostgres::Postgres(),
  dbname = "nemo",
  user = "orcabus"
)
d <- dtw_Wf_cttsov2(
  path = path, prefix = prefix, outdir = outdir,
  format = "db",
  dryrun = F,
  drid = drid,
  dbconn = dbconn
)
}
}
