% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{dtw_Wf_dragen}
\alias{dtw_Wf_dragen}
\title{Wf_dragen Download Tidy and Write}
\usage{
dtw_Wf_dragen(
  path,
  prefix,
  outdir,
  outdir_tidy = file.path(outdir, "dracarys_tidy"),
  format = "rds",
  max_files = 1000,
  dryrun = FALSE
)
}
\arguments{
\item{path}{Path to directory with raw workflow results (S3 or local filesystem).}

\item{prefix}{The LibraryID prefix of the sample.}

\item{outdir}{Path to output directory with raw files.}

\item{outdir_tidy}{Path to output directory with tidy files.}

\item{format}{Format of output files.}

\item{max_files}{Max number of files to list.}

\item{dryrun}{If TRUE, just list the files that will be downloaded (don't
download them).}
}
\value{
Tibble of tidy tibbles.
}
\description{
Downloads files from the \code{dragen} workflow and writes them in a tidy format.
}
\examples{
\dontrun{
#---- Local ----#

#---- S3 ----#
path <- file.path(
  "s3://pipeline-prod-cache-503977275616-ap-southeast-2/byob-icav2/production",
  "analysis/wgts-qc/20241123ffa837c4/L2401621_dragen_alignment"
)
prefix <- "L2401621"
outdir <- sub("s3:/", "~/s3", path)
dragen_tidy <- dtw_Wf_dragen(
  path = path, prefix = prefix, outdir = outdir,
  format = "tsv",
  dryrun = F
)
}
}
