% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{dtw_Wf_tso_ctdna_tumor_only}
\alias{dtw_Wf_tso_ctdna_tumor_only}
\title{Wf_tso_ctdna_tumor_only Download Tidy and Write}
\usage{
dtw_Wf_tso_ctdna_tumor_only(
  path,
  prefix,
  outdir,
  outdir_tidy = file.path(outdir, "dracarys_tidy"),
  format = "rds",
  max_files = 1000,
  dryrun = FALSE
)
}
\arguments{
\item{path}{Path to directory with raw workflow results (from S3 or
local filesystem).}

\item{prefix}{The SubjectID_LibraryID prefix of the sample (needed for path lookup).}

\item{outdir}{Path to output directory with raw files.}

\item{outdir_tidy}{Path to output directory with tidy files.}

\item{format}{Format of output files.}

\item{max_files}{Max number of files to list.}

\item{dryrun}{If TRUE, just list the files that will be downloaded (don't
download them).}
}
\value{
Tibble of tidy tibbles.
}
\description{
Downloads files from the \code{tso_ctdna_tumor_only} workflow and writes them in a tidy format.
}
\examples{
\dontrun{
p <- file.path(
  "gds://production/analysis_data/SBJ05563/tso_ctdna_tumor_only",
  "20240914d41300cd/L2401388/Results"
)
prefix <- "PRJ241446_L2401388"
outdir <- file.path(sub("gds:/", "~/icav1/g", p))
d <- dtw_Wf_tso_ctdna_tumor_only(
  path = p, prefix = prefix, outdir = outdir,
  format = "tsv",
  dryrun = F
)
}
}
