% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsov2.R
\name{dtw_Wf_tso_ctdna_tumor_only_v2}
\alias{dtw_Wf_tso_ctdna_tumor_only_v2}
\title{Wf_tso_ctdna_tumor_only_v2 Download Tidy and Write}
\usage{
dtw_Wf_tso_ctdna_tumor_only_v2(
  path,
  prefix,
  outdir,
  outdir_tidy = file.path(outdir, "dracarys_tidy"),
  format = "rds",
  max_files = 1000,
  dryrun = FALSE
)
}
\arguments{
\item{path}{Path to directory with raw workflow results (S3 or
local filesystem).}

\item{prefix}{The LibraryID prefix of the sample (needed for path lookup).}

\item{outdir}{Path to output directory with raw files.}

\item{outdir_tidy}{Path to output directory with tidy files.}

\item{format}{Format of output files.}

\item{max_files}{Max number of files to list.}

\item{dryrun}{If TRUE, just list the files that will be downloaded (don't
download them).}
}
\value{
Tibble of tidy tibbles.
}
\description{
Downloads files from the \code{tso_ctdna_tumor_only_v2} workflow and writes them in a tidy format.
}
\examples{
\dontrun{
p <- file.path(
  "s3://pipeline-prod-cache-503977275616-ap-southeast-2/byob-icav2/production",
  "analysis/cttsov2/20240915ff0295ed"
)
prefix <- "L2401290"
outdir <- sub("s3:/", "~/s3", p)
d <- dtw_Wf_tso_ctdna_tumor_only_v2(
  path = p, prefix = prefix, outdir = outdir,
  format = "tsv",
  dryrun = F
)
}
}
