% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen_fastqc.R
\name{FastqcMetricsFile}
\alias{FastqcMetricsFile}
\title{FastqcMetrics R6 Class}
\description{
Contains methods for reading and displaying contents of
the \code{fastqc_metrics.csv} file output from DRAGEN.
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{FastqcMetricsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FastqcMetricsFile-read}{\code{FastqcMetricsFile$read()}}
\item \href{#method-FastqcMetricsFile-write}{\code{FastqcMetricsFile$write()}}
\item \href{#method-FastqcMetricsFile-clone}{\code{FastqcMetricsFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastqcMetricsFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-FastqcMetricsFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{fastqc_metrics.csv} file output from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastqcMetricsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble. TODO.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastqcMetricsFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-FastqcMetricsFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{fastqc_metrics.csv} file output
from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastqcMetricsFile$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FastqcMetricsFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FastqcMetricsFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FastqcMetricsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
