% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{file_regex_getter}
\alias{file_regex_getter}
\title{Get dracarys \code{DR_FILE_REGEX}}
\usage{
file_regex_getter()
}
\value{
\code{DR_FILE_REGEX} R tibble object.
}
\description{
Get dracarys \code{DR_FILE_REGEX}
}
