% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{FineHistFile}
\alias{FineHistFile}
\title{FineHistFile R6 Class}
\description{
Contains methods for reading and displaying contents of
the \verb{wgs_fine_hist_<phenotype>.csv} file output from DRAGEN.
This file contains two columns: Depth and Overall.
The value in the Depth column ranges from 0 to 1000+ and the Overall
column indicates the number of loci covered at the corresponding depth.
}
\examples{
x1 <- system.file("extdata/wgs/SEQC-II.wgs_fine_hist_normal.csv.gz", package = "dracarys")
x2 <- system.file("extdata/wgs/SEQC-II.wgs_fine_hist_tumor.csv.gz", package = "dracarys")
ch1 <- FineHistFile$new(x1)
ch2 <- FineHistFile$new(x2)
d1 <- read(ch1)
d2 <- read(ch2)
ch1$plot(d1)
ch2$plot(d2)
ch1$write(d1, out_dir = tempdir(), prefix = "seqc_n", out_format = "tsv")
ch2$write(d2, out_dir = tempdir(), prefix = "seqc_t", out_format = "tsv")
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{FineHistFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FineHistFile-read}{\code{FineHistFile$read()}}
\item \href{#method-FineHistFile-write}{\code{FineHistFile$write()}}
\item \href{#method-FineHistFile-plot}{\code{FineHistFile$plot()}}
\item \href{#method-FineHistFile-clone}{\code{FineHistFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineHistFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-FineHistFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \verb{wgs_fine_hist_<phenotype>.csv} file output from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineHistFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with three columns:
\itemize{
\item label
\item depth
\item number of loci with given depth
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineHistFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-FineHistFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \verb{wgs_fine_hist_<phenotype>.csv} file output
from DRAGEN
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineHistFile$write(d, out_dir = NULL, prefix, out_format = "tsv", drid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineHistFile-plot"></a>}}
\if{latex}{\out{\hypertarget{method-FineHistFile-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the \verb{wgs_fine_hist_<phenotype>.csv} files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineHistFile$plot(d, x_lim = c(0, 300))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{x_lim}}{X axis range to plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object with depth of coverage on X axis,
and number of loci with that depth on Y axis.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FineHistFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FineHistFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineHistFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
