% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{gds_file_download}
\alias{gds_file_download}
\title{GDS File Download}
\usage{
gds_file_download(gds, out, token = Sys.getenv("ICA_ACCESS_TOKEN"))
}
\arguments{
\item{gds}{Full path to GDS file.}

\item{out}{Path to output file.}

\item{token}{ICA access token (by default uses $ICA_ACCESS_TOKEN env var).}
}
\description{
GDS File Download
}
