% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{gds_file_download_api}
\alias{gds_file_download_api}
\title{GDS File Download via API}
\usage{
gds_file_download_api(
  gds_fileid,
  out_file,
  token = Sys.getenv("ICA_ACCESS_TOKEN")
)
}
\arguments{
\item{gds_fileid}{GDS file ID.}

\item{out_file}{Path to output file.}

\item{token}{ICA access token (by default uses $ICA_ACCESS_TOKEN env var).}
}
\description{
GDS File Download via API
}
\examples{
\dontrun{
gds_fileid <- "fil.f9aa2ba7af0c4330095d08dadd2e16b0"
out <- tempfile()
token <- Sys.getenv("ICA_ACCESS_TOKEN")
}
}
