% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_icav1.R
\name{gds_file_download_cli}
\alias{gds_file_download_cli}
\title{GDS File Download via CLI}
\usage{
gds_file_download_cli(gds, out, token = Sys.getenv("ICA_ACCESS_TOKEN"))
}
\arguments{
\item{gds}{Full path to GDS file.}

\item{out}{Path to output file.}

\item{token}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}
}
\description{
GDS File Download via CLI
}
