% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_icav1.R
\name{gds_file_presignedurl}
\alias{gds_file_presignedurl}
\title{GDS File Presigned URL}
\usage{
gds_file_presignedurl(gds_fileid, token)
}
\arguments{
\item{gds_fileid}{GDS file ID.}

\item{token}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}
}
\value{
Presigned URL if valid.
}
\description{
Returns presigned URL of given GDS file.
}
