% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{gds_files_list}
\alias{gds_files_list}
\title{GDS Files List}
\usage{
gds_files_list(gdsdir, token, page_size = 100, include = NULL)
}
\arguments{
\item{gdsdir}{Full path to GDS directory.}

\item{token}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}

\item{page_size}{Page size (def: 100).}

\item{include}{Use PresignedUrl to include presigned URLs to all files within
the GDS directory.}
}
\value{
Tibble with file basename, file size, file full data path, file dir name.
}
\description{
List files on ICA GDS filesystem.
}
