% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{gds_files_list_fastq}
\alias{gds_files_list_fastq}
\title{List FASTQs In GDS Directory}
\usage{
gds_files_list_fastq(gdsdir, token, include_url = FALSE, page_size = 100)
}
\arguments{
\item{gdsdir}{GDS directory.}

\item{token}{ICA access token.}

\item{include_url}{Include presigned URLs to all files within the GDS directory.}

\item{page_size}{Page size.}
}
\value{
A tibble with type, bname, size, file_id, path, and presigned URL.
}
\description{
List FASTQs In GDS Directory
}
\examples{
\dontrun{
prim <- "gds://production/primary_data"
run <- "240719_A00130_0323_BHMCYHDSXC/202407205bad380d/BiModal_BM-5L"
gdsdir <- file.path(prim, run)
token <- Sys.getenv("ICA_ACCESS_TOKEN")
include_url <- F
page_size <- 100
gds_files_list_fastq(gdsdir, token, include_url, page_size)
}
}
