% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{gds_files_list_filter_relevant}
\alias{gds_files_list_filter_relevant}
\title{List Relevant Files In GDS Directory}
\usage{
gds_files_list_filter_relevant(
  gdsdir,
  token = Sys.getenv("ICA_ACCESS_TOKEN"),
  pattern = NULL,
  include_url = FALSE,
  page_size = 100,
  regexes = DR_FILE_REGEX,
  ...
)
}
\arguments{
\item{gdsdir}{GDS directory.}

\item{token}{ICA access token.}

\item{pattern}{Pattern to further filter the returned file type tibble.}

\item{include_url}{Include presigned URLs to all files within the GDS directory (def: FALSE).}

\item{page_size}{Page size (def: 100).}

\item{regexes}{Tibble with \code{regex} and \code{fun}ction name.}
}
\value{
A tibble with type, bname, size, file_id, path, and presigned URL.
\dontrun{
gdsdir <- "gds://production/analysis_data/SBJ01155/umccrise/202408300c218043/L2101566__L2101565"
gds_files_list_filter_relevant(gdsdir)
}
}
\description{
Lists relevant files in a GDS directory.
}
