% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{gds_volumes_list}
\alias{gds_volumes_list}
\title{List GDS Volumes}
\usage{
gds_volumes_list(token, page_size = 10)
}
\arguments{
\item{token}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}

\item{page_size}{Page size (def: 10).}
}
\value{
A tibble with vol name and vol id.
}
\description{
Lists GDS volumes accessible by the provided ICA token.
}
