% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grep_file}
\alias{grep_file}
\title{Grep File Pattern}
\usage{
grep_file(path = ".", regexp)
}
\arguments{
\item{path}{Path to look for file.}

\item{regexp}{A regular expression (e.g. \link{.}csv$) passed on to \code{grep()} to filter paths.}
}
\value{
The path to the file or an empty string if no match is found.
}
\description{
Grep File Pattern
}
