% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{ica_token_validate}
\alias{ica_token_validate}
\title{Validate ICA access token}
\usage{
ica_token_validate(token = Sys.getenv("ICA_ACCESS_TOKEN"))
}
\arguments{
\item{token}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}
}
\value{
Returns the token if valid, or else errors out.
}
\description{
Validates ICA access token by parsing it and checking its expiration date.
}
