% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_files_list_filter_relevant}
\alias{local_files_list_filter_relevant}
\title{List Relevant Files In Local Directory}
\usage{
local_files_list_filter_relevant(path, regexes = DR_FILE_REGEX)
}
\arguments{
\item{path}{Path to local directory.}

\item{regexes}{Tibble with \code{regex} and \code{fun}ction name.}
}
\value{
A tibble with type, bname, size, file_id, path, and presigned URL.
}
\description{
Lists relevant files in a local directory.
}
\examples{
\dontrun{
path <- "~/icav1/g/production/analysis_data/SBJ01155/umccrise"
local_files_list_filter_relevant(path, regexes = DR_FILE_REGEX)
}
}
