% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_local.R
\name{local_list_files_dir}
\alias{local_list_files_dir}
\title{List Files in Local Directory}
\usage{
local_list_files_dir(localdir)
}
\arguments{
\item{localdir}{Path to local directory.}
}
\value{
A tibble with file basename, size, last modification timestamp
and full path.
}
\description{
Lists files in a local directory.
}
\examples{
localdir <- system.file("R", package = "dracarys")
x <- local_list_files_dir(localdir)
}
