% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_list_files_filter_relevant}
\alias{local_list_files_filter_relevant}
\title{List Relevant Files In Local Directory}
\usage{
local_list_files_filter_relevant(path, regexes = DR_FILE_REGEX)
}
\arguments{
\item{path}{Path to local directory.}

\item{regexes}{Tibble with \code{regex} and \code{fun}ction name (see example).}
}
\value{
A tibble with file type, basename, size, last modified timestamp, and
path.
}
\description{
Lists relevant files in a local directory.
}
\examples{
path <- system.file("extdata/tso", package = "dracarys")
regexes <- tibble::tibble(regex = "multiqc_data\\\\.json$", fun = "MultiqcFile")
x <- local_list_files_filter_relevant(path, regexes)
}
