% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MappingMetricsFile.R
\name{MappingMetricsFile}
\alias{MappingMetricsFile}
\title{MappingMetricsFile R6 Class}
\description{
Contains methods for reading and displaying contents of
the \code{mapping_metrics.csv} file output from DRAGEN.
This file contains mapping and aligning metrics, like the metrics computed by
the Samtools Flagstat command. These metrics are available on an aggregate
level (over all input data), and on a per read group level.
Unless explicitly stated, the metrics units are in reads (i.e., not in
terms of pairs or alignments).
}
\examples{
x <- system.file("extdata/wgs/SEQC-II.mapping_metrics.csv.gz", package = "dracarys")
mm <- MappingMetricsFile$new(x)
mm$read() # or read(mm)

}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{MappingMetricsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-read}{\code{MappingMetricsFile$read()}}
\item \href{#method-clone}{\code{MappingMetricsFile$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname">}\href{../../dracarys/html/File.html#method-bname}{\code{dracarys::File$bname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize">}\href{../../dracarys/html/File.html#method-initialize}{\code{dracarys::File$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print">}\href{../../dracarys/html/File.html#method-print}{\code{dracarys::File$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type">}\href{../../dracarys/html/File.html#method-type}{\code{dracarys::File$type()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{mapping_metrics.csv} file output from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappingMetricsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
- category: summary or read group
- Phenotype: e.g. tumor, normal
- RG: read group
- var: metric variable
- var_abbrev: metric variable abbreviation
- count: count of reads
- pct: percentage of reads
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MappingMetricsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
