% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{match_regex}
\alias{match_regex}
\title{Match File to Regex}
\usage{
match_regex(x, regexes = DR_FILE_REGEX)
}
\arguments{
\item{x}{File to match.}

\item{regexes}{Tibble with \code{regex} and \code{fun}ction name.}
}
\value{
The function corresponding to the matching regex from \code{DR_FILE_REGEX}, or
NA if there is no match made.
}
\description{
Matches a given file with the regexes found in \code{DR_FILE_REGEX} and if there is
a match, it returns the 'name' of that match.
}
\examples{
match_regex("foo.msi.json.gz")
match_regex("foo.fake.tsv")

}
