% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal_meta.R
\name{meta_tso_ctdna_tumor_only}
\alias{meta_tso_ctdna_tumor_only}
\title{Metadata for tso_ctdna_tumor_only workflow}
\usage{
meta_tso_ctdna_tumor_only(pmeta, status = c("Succeeded"))
}
\arguments{
\item{pmeta}{Path to portal workflows metadata table, or tibble with already parsed data.}

\item{status}{Workflow status to keep (default: Succeeded).}
}
\value{
A tibble with metadata per cttso workflow run.
}
\description{
Metadata for tso_ctdna_tumor_only workflow
}
\examples{
\dontrun{
pmeta <- here::here("nogit/data_portal/2023-05-07_workflows_5f43da12-0b6a-41ce-86c9-9ee62df8792e.csv")
cttso_metadata(pmeta)
}
}
