% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_column_map_append}
\alias{multiqc_column_map_append}
\title{Append New MultiQC Workflow Columns}
\usage{
multiqc_column_map_append(d, w, m)
}
\arguments{
\item{d}{Tidy MultiQC tibble with raw names (i.e. pre-rename).}

\item{w}{New workflow name.}

\item{m}{Path to the 'inst/extdata/multiqc_column_map.tsv' dracarys file.}
}
\description{
Append New MultiQC Workflow Columns
}
