% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_date_fmt}
\alias{multiqc_date_fmt}
\title{Format MultiQC Config Date}
\usage{
multiqc_date_fmt(cdate)
}
\arguments{
\item{cdate}{String with config date in "YYYY-MM-DD, HH:MM UTC" format.}
}
\value{
Properly formatted datetime string in "YYYY-MM-DDTHH:MM:SS" format.
If the input string was formatted differently, it will return NA.
}
\description{
Format MultiQC Config Date
}
\examples{
cdate <- "2023-04-07, 09:09 UTC"
(res1 <- multiqc_date_fmt(cdate))
(res2 <- multiqc_date_fmt("2023-04-07"))
(res3 <- multiqc_date_fmt("UNKNOWN"))
}
