% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_parse_gen}
\alias{multiqc_parse_gen}
\title{Parse MultiQC 'report_general_stats_data' JSON Element}
\usage{
multiqc_parse_gen(p)
}
\arguments{
\item{p}{Parsed MultiQC JSON.}
}
\value{
A list.
}
\description{
Parses MultiQC 'report_general_stats_data' JSON Element. Modified from the
awesome \url{https://github.com/multimeric/TidyMultiqc}.
}
