% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_parse_plots}
\alias{multiqc_parse_plots}
\title{Parse Plot Data from MultiQC JSON}
\usage{
multiqc_parse_plots(j, plot_names = NULL)
}
\arguments{
\item{j}{Path to \code{multiqc_data.json} file.}

\item{plot_names}{Names of plots to parse. Use "everything" if you wantz all
the plotz.}
}
\value{
Nested tibble with plot name and result as list column (use \code{tidyr::unnest} to access).
}
\description{
Parse Plot Data from MultiQC JSON
}
\examples{
\dontrun{
j <- here::here("nogit/warehouse/cttso/2023/05/wfr.ec3748db414f422381e419d367f73eec/dracarys_gds_sync/multiqc_data.json")
multiqc_parse_plots(j, plot_names = c("fastqc_gc_content_mean_sequence_quality_plot", "dragen_coverage_per_contig"))
multiqc_parse_plots(j, plot_names = NULL)
}
}
