% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_parse_raw}
\alias{multiqc_parse_raw}
\title{Parse MultiQC 'report_saved_raw_data' JSON Element}
\usage{
multiqc_parse_raw(p)
}
\arguments{
\item{p}{Parsed MultiQC JSON.}
}
\value{
A list.
}
\description{
Parses MultiQC 'report_saved_raw_data' JSON Element.
}
