% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_parse_raw_interop}
\alias{multiqc_parse_raw_interop}
\title{Parse Interop MultiQC 'report_saved_raw_data' JSON Element}
\usage{
multiqc_parse_raw_interop(p)
}
\arguments{
\item{p}{Parsed Interop MultiQC JSON.}
}
\value{
A list with summary and per-lane tibbles.
}
\description{
Parses Interop MultiQC 'report_saved_raw_data' JSON Element.
}
