% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_parse_xyline_plot}
\alias{multiqc_parse_xyline_plot}
\title{MultiQC Extract XY Line Plot Data}
\usage{
multiqc_parse_xyline_plot(dat)
}
\arguments{
\item{dat}{Parsed JSON list element with specific plot data to extract.}
}
\value{
Tibble with name and data list-column.
}
\description{
Extracts \code{xy_line} data for the given plot name from a MultiQC JSON.
}
\details{
Note that the \code{dragen_coverage_per_contig/non_main_contig} plots do not conform
to the normal structure of \code{xy_line} plots, so we handle those separately.
}
