% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_parse_xyline_plot_contig_cvg}
\alias{multiqc_parse_xyline_plot_contig_cvg}
\title{MultiQC Extract XY Line Plot Data for DRAGEN Contig Coverage}
\usage{
multiqc_parse_xyline_plot_contig_cvg(dat)
}
\arguments{
\item{dat}{Parsed JSON list element with specific plot data to extract.}
}
\value{
Tibble with name and data list-column.
}
\description{
MultiQC Extract XY Line Plot Data for DRAGEN Contig Coverage
}
