% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_tidy_json}
\alias{multiqc_tidy_json}
\title{Tidy MultiQC JSON}
\usage{
multiqc_tidy_json(j)
}
\arguments{
\item{j}{Path to \code{multiqc_data.json}.}
}
\value{
A tidy tibble where each column corresponds to a single metric,
and each row corresponds to a single sample.
}
\description{
Tidies 'multiqc_data.json' output from MultiQC.
Modified from the awesome \url{https://github.com/multimeric/TidyMultiqc}.
}
