% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgr.R
\name{pcgr_json_read}
\alias{pcgr_json_read}
\title{PCGR JSON Read}
\usage{
pcgr_json_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
A tibble with: \code{fracIndels}, \code{predicted_class}, \code{tmb_estimate}, \code{n_tmb}.
}
\description{
PCGR JSON Read
}
\examples{
\dontrun{
pcgr_json_read(x)
}
}
