% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgr.R
\name{PcgrJsonFile}
\alias{PcgrJsonFile}
\title{PcgrJson R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{pcgr.json.gz} file output from PCGR.
}
\examples{
\dontrun{
x <- "/path/to/pcgr.json.gz"
d <- PcgrJsonFile$new(x)
d_parsed <- d$read() # or read(d)
d$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "both")
}
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{PcgrJsonFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PcgrJsonFile-read}{\code{PcgrJsonFile$read()}}
\item \href{#method-PcgrJsonFile-write}{\code{PcgrJsonFile$write()}}
\item \href{#method-PcgrJsonFile-clone}{\code{PcgrJsonFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PcgrJsonFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-PcgrJsonFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{pcgr.json.gz} file output from PCGR.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PcgrJsonFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of tibbles.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PcgrJsonFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-PcgrJsonFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{pcgr.json.gz} file output from PCGR.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PcgrJsonFile$write(d, out_dir, prefix, out_format = "tsv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s) (one of 'tsv' (def.),
'parquet', 'both').}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PcgrJsonFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PcgrJsonFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PcgrJsonFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
