% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgr.R
\name{PcgrTiersFile}
\alias{PcgrTiersFile}
\title{PcgrTiersFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{pcgr.snvs_indels.tiers.tsv} file output from PCGR.
}
\examples{
\dontrun{
x <- "/path/to/pcgr.snvs_indels.tiers.tsv"
d <- PcgrTiersFile$new(x)
d_parsed <- d$read() # or read(d)
d$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "both")
}
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{PcgrTiersFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PcgrTiersFile-read}{\code{PcgrTiersFile$read()}}
\item \href{#method-PcgrTiersFile-write}{\code{PcgrTiersFile$write()}}
\item \href{#method-PcgrTiersFile-clone}{\code{PcgrTiersFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PcgrTiersFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-PcgrTiersFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{pcgr.snvs_indels.tiers.tsv} file output from PCGR.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PcgrTiersFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List of tibbles.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PcgrTiersFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-PcgrTiersFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{pcgr.snvs_indels.tiers.tsv} file output from PCGR.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PcgrTiersFile$write(d, out_dir, prefix, out_format = "tsv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s) (one of 'tsv' (def.),
'parquet', 'both').}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PcgrTiersFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PcgrTiersFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PcgrTiersFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
