% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PloidyEstimationMetrics.R
\name{PloidyEstimationMetrics}
\alias{PloidyEstimationMetrics}
\title{PloidyEstimationMetrics R6 Class}
\description{
Contains methods for reading contents of
the \code{ploidy_estimation_metrics.csv} file output from DRAGEN.
}
\examples{
x <- system.file("extdata/SEQC-II.ploidy_estimation_metrics.csv", package = "dracarys")
pem <- PloidyEstimationMetrics$new(x)
pem$read() # or read(pem)

}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{PloidyEstimationMetrics}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-read}{\code{PloidyEstimationMetrics$read()}}
\item \href{#method-clone}{\code{PloidyEstimationMetrics$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname">}\href{../../dracarys/html/File.html#method-bname}{\code{dracarys::File$bname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize">}\href{../../dracarys/html/File.html#method-initialize}{\code{dracarys::File$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print">}\href{../../dracarys/html/File.html#method-print}{\code{dracarys::File$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type">}\href{../../dracarys/html/File.html#method-type}{\code{dracarys::File$type()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{ploidy_estimation_metrics.csv} file output from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PloidyEstimationMetrics$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
- label: sample label (inferred from file name)
- var: variable of interest (e.g. X median coverage)
- value: value for specific variable (e.g. X median coverage
variable with a value of  50)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PloidyEstimationMetrics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
