% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{read.File}
\alias{read.File}
\title{Read File object}
\usage{
\method{read}{File}(x, ...)
}
\arguments{
\item{x}{Object of class \code{File}.}

\item{...}{Arguments passed on to appropriate read_* function.}
}
\description{
Read the file based on its type.
}
