% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Generic Reading}
\usage{
read(x, ...)
}
\arguments{
\item{x}{Object of the respective class.}

\item{...}{Additional arguments.}
}
\value{
Possibly a tibble or a list, depending on the object.
}
\description{
Generic function for reading dracarys objects.
}
