% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{ReplayFile}
\alias{ReplayFile}
\title{ReplayFile R6 Class}
\description{
Contains methods for reading contents of
the \code{replay.json} file output from DRAGEN, which contains the DRAGEN command
line, parameters and version for the specific run.
}
\examples{
x <- system.file("extdata/wgs/SEQC-II-replay.json.gz", package = "dracarys")
r <- ReplayFile$new(x)
d <- r$read() # or read(r)
r$write(d, out_dir = tempdir(), prefix = "seqc_replay", out_format = "tsv")
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{ReplayFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReplayFile-read}{\code{ReplayFile$read()}}
\item \href{#method-ReplayFile-write}{\code{ReplayFile$write()}}
\item \href{#method-ReplayFile-clone}{\code{ReplayFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplayFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-ReplayFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{replay.json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplayFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with one row and metrics spread across individual columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplayFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-ReplayFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{replay.json} file output
from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplayFile$write(d, out_dir = NULL, prefix, out_format = "tsv", drid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReplayFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReplayFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplayFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
