% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_files_list_filter_relevant}
\alias{s3_files_list_filter_relevant}
\title{List Relevant Files In AWS S3 Directory}
\usage{
s3_files_list_filter_relevant(
  s3dir,
  pattern = NULL,
  page_size = 1000,
  max_items = 1000,
  presign = FALSE,
  expiry_sec = 43200
)
}
\arguments{
\item{s3dir}{GDS directory.}

\item{pattern}{Pattern to further filter the returned file type tibble.}

\item{page_size}{The size of each page to get in the AWS service call (def: 1000).}

\item{max_items}{The total number of items to return in the command’s output (def: 1000).}

\item{presign}{Include presigned URLs (def: FALSE).}

\item{expiry_sec}{Number of seconds the presigned URL will be valid for (if generated) (def: 43200 (12hrs)).}
}
\value{
A tibble with path, date, file size, file type, and presigned URL if requested.
}
\description{
Lists relevant files in an AWS S3 directory.
}
\examples{
\dontrun{
s3dir <- "s3://umccr-primary-data-prod/cancer_report_tables"
s3_files_list_filter_relevant(s3dir = s3dir, presign = TRUE)
}
}
