% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_s3.R
\name{s3_search}
\alias{s3_search}
\title{Search AWS S3 Objects}
\usage{
s3_search(pat, rows)
}
\arguments{
\item{pat}{Pattern to search for (e.g. 'multiqc_data.json').}

\item{rows}{Max number of rows to return.}
}
\value{
Tibble with S3 path, object size, date modified, id, unique hash.
}
\description{
Searches for the given pattern in the UMCCR \code{umccr-primary-data-prod} AWS S3
bucket.
}
\examples{
\dontrun{
pat <- "qc_summary.tsv.gz"
s3_search(pat, 10)
}
}
