% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_s3.R
\name{s3_sync}
\alias{s3_sync}
\title{AWS S3 Sync Helper}
\usage{
s3_sync(src, dest, pats = NULL)
}
\arguments{
\item{src}{(\code{character(1)})\cr
S3 source path.}

\item{dest}{(\code{character(1)})\cr
Local destination path.}

\item{pats}{(\code{tibble()})\cr
Patterns tibble with \code{inex} ("in" or "ex") and \code{pat} (pattern) columns.}
}
\description{
AWS S3 Sync Helper
}
\examples{
\dontrun{
src <- "s3://my-awesome-bucket/path/to/run1"
dest <- sub("s3:/", "~/s3", src)
s3_sync(src, dest)
}
}
