% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{schema2django}
\alias{schema2django}
\title{Generate Django Models from a Schema Tibble}
\usage{
schema2django(schema, out)
}
\arguments{
\item{schema_tbl}{A tibble containing schema information (from \code{s1$schema()})}

\item{output_file}{Name of the Python file to write the Django models (default: "models.py")}
}
\value{
Writes Django models to a file
}
\description{
Generate Django Models from a Schema Tibble
}
\examples{
schema <- tibble::tibble(
  table_name = "foo",
  column_name = c("dracarysId", "foo_num", "foo_dbl", "foo_chr", "foo_int", "foo_date"),
  data_type = c("TextField", "FloatField", "FloatField", "TextField", "IntegerField", "DateField"),
  primary_key = rep(c(TRUE, FALSE), c(1L, 5L)),
)
out <- tempfile()
schema2django(s1$schema(), out)
}
