% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TblSchema.R
\name{TblSchema}
\alias{TblSchema}
\title{TblSchema R6 Class}
\description{
Encapsulates a tibble with metadata such as column types
and primary keys. It allows schema inference, and export to YAML for
integration with Django.
}
\examples{
tbl <- tibble::tibble(
  dracarysId = "abcd1234",
  foo_num = 123,
  foo_dbl = 3.14,
  foo_chr = "foobar",
  foo_int = 35L,
  foo_date = Sys.Date()
)
s1 <- TblSchema$new(tbl = tbl, name = "Foo", pk = "dracarysId")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TblSchema-new}{\code{TblSchema$new()}}
\item \href{#method-TblSchema-schema}{\code{TblSchema$schema()}}
\item \href{#method-TblSchema-print}{\code{TblSchema$print()}}
\item \href{#method-TblSchema-clone}{\code{TblSchema$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TblSchema-new"></a>}}
\if{latex}{\out{\hypertarget{method-TblSchema-new}{}}}
\subsection{Method \code{new()}}{
Create a new TblSchema object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TblSchema$new(tbl = NULL, name = NULL, pk = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl}}{Tibble object.}

\item{\code{name}}{Name of table.}

\item{\code{pk}}{Character vector of primary key(s).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TblSchema-schema"></a>}}
\if{latex}{\out{\hypertarget{method-TblSchema-schema}{}}}
\subsection{Method \code{schema()}}{
Schema of tbl.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TblSchema$schema()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TblSchema-print"></a>}}
\if{latex}{\out{\hypertarget{method-TblSchema-print}{}}}
\subsection{Method \code{print()}}{
Print details about the File.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TblSchema$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TblSchema-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TblSchema-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TblSchema$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
