% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_files}
\alias{tidy_files}
\title{Tidy Files}
\usage{
tidy_files(x, envir = parent.frame())
}
\arguments{
\item{x}{Tibble with \code{localpath} to file and the function \code{type} to parse it.}

\item{envir}{the environment in which to evaluate the function e.g. use \code{self}
when using inside R6 classes.}
}
\value{
Tibble with parsed data in a \code{data} list-column.
}
\description{
Tidy Files
}
\examples{
\dontrun{
p1 <- "~/icav1/g/production/analysis_data/SBJ01155/umccrise/202408300c218043"
p2 <- "L2101566__L2101565/SBJ01155__PRJ211091-qc_summary.tsv.gz"
p <- file.path(p1, p2)
x <- tibble::tibble(type = "readr::read_tsv", localpath = p)
tidy_files(x)
)
}

}
