% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{time_metrics_process}
\alias{time_metrics_process}
\title{Process Multiple TimeMetricsFile Objects}
\usage{
time_metrics_process(x, id = seq_len(length(x)))
}
\arguments{
\item{x}{Atomic vector with one or more TimeMetricsFile objects.}

\item{id}{ID for each input, which is used to disambiguate files
generated from same samples. Default: index from 1 to length of \code{x}.}
}
\value{
tibble with the following columns:
\itemize{
\item Step: DRAGEN step
\item Time: time in HH:MM
}
}
\description{
Processes multiple TimeMetricsFile objects.
}
\examples{
p <- system.file("extdata/wgs/SEQC-II.time_metrics.csv.gz", package = "dracarys")
x <- TimeMetricsFile$new(p)
(tm <- time_metrics_process(c(x, x), id = c("run1", "run2")))

}
