% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso_acfc.R
\name{tso_acfc_plot}
\alias{tso_acfc_plot}
\title{Plot ACFC}
\usage{
tso_acfc_plot(d, max_num = 15)
}
\arguments{
\item{d}{Parsed object from \code{tso_acfc_read}.}

\item{max_num}{Maximum number to display in both plots.}
}
\value{
Both histogram plot objects.
}
\description{
Generates the UmiStatistics Histogram plots from the
\code{AlignCollapseFusionCaller_metrics.json.gz} file output from TSO.
}
\details{
\itemize{
\item Histo is the majority from UmiStatistics section, deal with it separately.
\item Histo of num supporting fragments: Num of families with 0/1/2/3... raw reads.
\item Histo of unique UMIs per fragment pos: Num of pos with 0/1/2/3... UMI seqs.
}
}
