% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso_acfc.R
\name{tso_acfc_read}
\alias{tso_acfc_read}
\title{TsoAlignCollapseFusionCallerMetricsFile R6 Class}
\usage{
tso_acfc_read(x)
}
\description{
Contains methods for reading and displaying contents of the
\code{AlignCollapseFusionCaller_metrics.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705.AlignCollapseFusionCaller_metrics.json.gz",
  package = "dracarys"
)
d <- tso_acfc_read(x)
p <- tso_acfc_plot(d)
}
