% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{tso_combinedvaro_smallv_read}
\alias{tso_combinedvaro_smallv_read}
\title{Read TSO CombinedVariantOutput File}
\usage{
tso_combinedvaro_smallv_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\description{
Reads \code{CombinedVariantOutput.tsv} output from the TSO500 workflow and extracts
only the Small Variants section (due to inconsistencies with other sections).
}
