% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{tso_msi_read}
\alias{tso_msi_read}
\title{Read TSO MSI JSON File}
\usage{
tso_msi_read(x)
}
\description{
Reads \code{msi.json.gz} file output from the TSO500 workflow.
}
\examples{
x <- system.file("extdata/tso/sample705.msi.json.gz", package = "dracarys")
tso_msi_read(x)
}
