% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{tso_rmd}
\alias{tso_rmd}
\title{Generate TSO Report}
\usage{
tso_rmd(prefix = NULL, out_file = "tso_report.html", quiet = FALSE)
}
\arguments{
\item{prefix}{Prefix for input files generated from UMCCR TSO500 ctDNA workflow (i.e. '/path/to/sbjx'.ext, where 'sbjx'
is the common prefix of all relevant files in directory '/path/to/').}

\item{out_file}{Name of output HTML file (needs '.html' suffix) (def: \code{tso_report.html}).}

\item{quiet}{Suppress warning and other messages.}
}
\value{
Path to rendered HTML report.
}
\description{
Generates a HTML report with results from the UMCCR TSO500 ctDNA workflow.
}
