% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso_sar.R
\name{tso_sar_read}
\alias{tso_sar_read}
\title{Read TSO SampleAnalysisResults JSON File}
\usage{
tso_sar_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\value{
list of tibbles
}
\description{
Contains methods for reading and displaying contents of the
\code{SampleAnalysisResults.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705_SampleAnalysisResults.json.gz", package = "dracarys")
d <- tso_sar_read(x)
options(pillar.max_dec_width = 99) # for scientific notation switchoff
d |>
  dplyr::filter(name == "sar_qc") |>
  dplyr::select(data) |>
  tidyr::unnest(data) |>
  tidyr::pivot_longer(dplyr::everything())
}
