% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{tso_targetregcvg_read}
\alias{tso_targetregcvg_read}
\title{Read TSO TargetRegionCoverage File}
\usage{
tso_targetregcvg_read(x)
}
\value{
tibble with the following columns:
\itemize{
\item ConsensusReadDepth
\item BasePair
\item Percentage
}
}
\description{
Reads the \code{TargetRegionCoverage.json.gz} file output from the TSO500 workflow.
}
\examples{
x <- system.file("extdata/tso/sample705.TargetRegionCoverage.json.gz", package = "dracarys")
d <- tso_targetregcvg_read(x)
tso_targetregcvg_plot(d, min_pct = 0)
}
