% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{tso_tmbt_read}
\alias{tso_tmbt_read}
\title{Read TSO TMB_Trace File}
\usage{
tso_tmbt_read(x)
}
\description{
Reads the \code{TMB_Trace.tsv} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705_TMB_Trace.tsv", package = "dracarys")
tso_tmbt_read(x)
}
