% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoCombinedVariantOutputFile}
\alias{TsoCombinedVariantOutputFile}
\title{TsoCombinedVariantOutputFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{CombinedVariantOutput.tsv} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705_CombinedVariantOutput.tsv", package = "dracarys")
d <- TsoCombinedVariantOutputFile$new(x)
d$read() # or read(d)
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoCombinedVariantOutputFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoCombinedVariantOutputFile-read}{\code{TsoCombinedVariantOutputFile$read()}}
\item \href{#method-TsoCombinedVariantOutputFile-clone}{\code{TsoCombinedVariantOutputFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoCombinedVariantOutputFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoCombinedVariantOutputFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{CombinedVariantOutput.tsv} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoCombinedVariantOutputFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of following tibbles:
\itemize{
\item FragmentLength
\item Count
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoCombinedVariantOutputFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoCombinedVariantOutputFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoCombinedVariantOutputFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
