% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoCopyNumberVariantsVcfFile}
\alias{TsoCopyNumberVariantsVcfFile}
\title{TsoCopyNumberVariantsVcfFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{CopyNumberVariants.vcf.gz} file output from TSO.
}
\examples{
\dontrun{
x <- normalizePath("~/icav1/g/production/analysis_data/SBJ00704/tso_ctdna_tumor_only/202311242b9666e0/L2301411/Results/PTC_ctTSO231120_L2301411/dracarys_gds_sync/PTC_ctTSO231120_L2301411_CopyNumberVariants.vcf.gz")
d <- TsoCopyNumberVariantsVcfFile$new(x)
d_parsed <- d$read() # or read(d)
d$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "tsv")
d$write(d_parsed, prefix = "FOO", out_format = "delta", drid = "wfr.123")
}
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoCopyNumberVariantsVcfFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoCopyNumberVariantsVcfFile-read}{\code{TsoCopyNumberVariantsVcfFile$read()}}
\item \href{#method-TsoCopyNumberVariantsVcfFile-write}{\code{TsoCopyNumberVariantsVcfFile$write()}}
\item \href{#method-TsoCopyNumberVariantsVcfFile-clone}{\code{TsoCopyNumberVariantsVcfFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoCopyNumberVariantsVcfFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoCopyNumberVariantsVcfFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{CopyNumberVariants.vcf.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoCopyNumberVariantsVcfFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with variants.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoCopyNumberVariantsVcfFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TsoCopyNumberVariantsVcfFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{CopyNumberVariants.vcf.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoCopyNumberVariantsVcfFile$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoCopyNumberVariantsVcfFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoCopyNumberVariantsVcfFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoCopyNumberVariantsVcfFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
