% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoFragmentLengthHistFile}
\alias{TsoFragmentLengthHistFile}
\title{TsoFragmentLengthHistFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{fragment_length_hist.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705.fragment_length_hist.json.gz", package = "dracarys")
fl <- TsoFragmentLengthHistFile$new(x)
fl$read() # or read(fl)
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoFragmentLengthHistFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoFragmentLengthHistFile-read}{\code{TsoFragmentLengthHistFile$read()}}
\item \href{#method-TsoFragmentLengthHistFile-clone}{\code{TsoFragmentLengthHistFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFragmentLengthHistFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFragmentLengthHistFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{fragment_length_hist.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFragmentLengthHistFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item FragmentLength
\item Count
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFragmentLengthHistFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFragmentLengthHistFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFragmentLengthHistFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
