% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoMergedSmallVariantsVcfFile}
\alias{TsoMergedSmallVariantsVcfFile}
\title{TsoMergedSmallVariantsVcfFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{MergedSmallVariants.vcf.gz} file output from TSO.
}
\examples{
\dontrun{
x <- here::here("nogit/tso/2023-05-30/SBJ00595_L2100178/PTC_SrSqCMM1pc_L2100178_rerun_MergedSmallVariants.vcf.gz")
d <- TsoMergedSmallVariantsVcfFile$new(x)
d_parsed <- d$read() # or read(d)
d$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "both")
d$write(d_parsed, prefix = "FOO", out_format = "delta", drid = "wfr.123")
}
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoMergedSmallVariantsVcfFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoMergedSmallVariantsVcfFile-read}{\code{TsoMergedSmallVariantsVcfFile$read()}}
\item \href{#method-TsoMergedSmallVariantsVcfFile-write}{\code{TsoMergedSmallVariantsVcfFile$write()}}
\item \href{#method-TsoMergedSmallVariantsVcfFile-clone}{\code{TsoMergedSmallVariantsVcfFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoMergedSmallVariantsVcfFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoMergedSmallVariantsVcfFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{MergedSmallVariants.vcf.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoMergedSmallVariantsVcfFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with variants.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoMergedSmallVariantsVcfFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TsoMergedSmallVariantsVcfFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{MergedSmallVariants.vcf.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoMergedSmallVariantsVcfFile$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory. Leave NULL if using on Databricks.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s) (one of 'tsv' (def.),
'parquet', 'both').}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoMergedSmallVariantsVcfFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoMergedSmallVariantsVcfFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoMergedSmallVariantsVcfFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
