% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoSampleAnalysisResultsFile}
\alias{TsoSampleAnalysisResultsFile}
\title{TsoSampleAnalysisResultsFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{SampleAnalysisResults.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705_SampleAnalysisResults.json.gz", package = "dracarys")
res <- TsoSampleAnalysisResultsFile$new(x)
res$read() # or read(res)
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoSampleAnalysisResultsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoSampleAnalysisResultsFile-read}{\code{TsoSampleAnalysisResultsFile$read()}}
\item \href{#method-TsoSampleAnalysisResultsFile-clone}{\code{TsoSampleAnalysisResultsFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoSampleAnalysisResultsFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoSampleAnalysisResultsFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{SampleAnalysisResults.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoSampleAnalysisResultsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item \code{TOTAL_PF_READS}
\item \code{MEAN_FAMILY_SIZE}
\item \code{MEDIAN_TARGET_COVERAGE}
\item \code{PCT_CHIMERIC_READS}
\item \code{PCT_EXON_500X}
\item \code{PCT_EXON_1500X}
\item \code{PCT_READ_ENRICHMENT}
\item \code{PCT_USABLE_UMI_READS}
\item \code{MEAN_TARGET_COVERAGE}
\item \code{PCT_ALIGNED_READS}
\item \code{PCT_CONTAMINATION_EST}
\item \code{PCT_TARGET_0.4X_MEAN}
\item \code{PCT_TARGET_500X}
\item \code{PCT_TARGET_1000X}
\item \code{PCT_TARGET_1500X}
\item \code{PCT_DUPLEXFAMILIES}
\item \code{MEDIAN_INSERT_SIZE}
\item \code{MAX_SOMATIC_AF}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoSampleAnalysisResultsFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoSampleAnalysisResultsFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoSampleAnalysisResultsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
