% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoTmbTraceTsvFile}
\alias{TsoTmbTraceTsvFile}
\title{TsoTmbTraceTsvFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{TMB_Trace.tsv} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705_TMB_Trace.tsv", package = "dracarys")
d <- TsoTmbTraceTsvFile$new(x)
d_parsed <- d$read() # or read(d)
d$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "both")
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoTmbTraceTsvFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoTmbTraceTsvFile-read}{\code{TsoTmbTraceTsvFile$read()}}
\item \href{#method-TsoTmbTraceTsvFile-write}{\code{TsoTmbTraceTsvFile$write()}}
\item \href{#method-TsoTmbTraceTsvFile-clone}{\code{TsoTmbTraceTsvFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoTmbTraceTsvFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoTmbTraceTsvFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{TMB_Trace.tsv} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoTmbTraceTsvFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item FragmentLength
\item Count
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoTmbTraceTsvFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TsoTmbTraceTsvFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{TMB_Trace.tsv} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoTmbTraceTsvFile$write(d, out_dir, prefix, out_format = "tsv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s) (one of 'tsv' (def.),
'parquet', 'both').}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoTmbTraceTsvFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoTmbTraceTsvFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoTmbTraceTsvFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
