% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umccrise.R
\name{UmccriseCanRepTables}
\alias{UmccriseCanRepTables}
\title{UmccriseCanRepTables R6 Class}
\description{
Reads and writes tidy versions of files within the \code{cancer_report_tables} directory
output from the \code{umccrise} workflow.
}
\examples{
\dontrun{
p1 <- "~/icav1/g/production/analysis_data/SBJ01155/umccrise/202408300c218043"
p2 <- "L2101566__L2101565/SBJ01155__PRJ211091/cancer_report_tables"
p <- file.path(p1, p2)
obj <- UmccriseCanRepTables$new(p)
obj$path
obj$contents
d <- obj$read()
obj$write(d, out_dir = tempdir(), prefix = "sampleA", out_format = "tsv")
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Path to the \code{cancer_report_tables} directory.}

\item{\code{contents}}{Tibble with file path, basename, and size.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UmccriseCanRepTables-new}{\code{UmccriseCanRepTables$new()}}
\item \href{#method-UmccriseCanRepTables-print}{\code{UmccriseCanRepTables$print()}}
\item \href{#method-UmccriseCanRepTables-grep_file}{\code{UmccriseCanRepTables$grep_file()}}
\item \href{#method-UmccriseCanRepTables-read_chordtsv}{\code{UmccriseCanRepTables$read_chordtsv()}}
\item \href{#method-UmccriseCanRepTables-read_hrdetecttsv}{\code{UmccriseCanRepTables$read_hrdetecttsv()}}
\item \href{#method-UmccriseCanRepTables-read_sigs}{\code{UmccriseCanRepTables$read_sigs()}}
\item \href{#method-UmccriseCanRepTables-read_qcsummarytsv}{\code{UmccriseCanRepTables$read_qcsummarytsv()}}
\item \href{#method-UmccriseCanRepTables-read}{\code{UmccriseCanRepTables$read()}}
\item \href{#method-UmccriseCanRepTables-write}{\code{UmccriseCanRepTables$write()}}
\item \href{#method-UmccriseCanRepTables-clone}{\code{UmccriseCanRepTables$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmccriseCanRepTables-new"></a>}}
\if{latex}{\out{\hypertarget{method-UmccriseCanRepTables-new}{}}}
\subsection{Method \code{new()}}{
Create a new UmccriseCanRepTables object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmccriseCanRepTables$new(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to the \code{cancer_report_tables} directory.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmccriseCanRepTables-print"></a>}}
\if{latex}{\out{\hypertarget{method-UmccriseCanRepTables-print}{}}}
\subsection{Method \code{print()}}{
Print details about the cancer_report_tables directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmccriseCanRepTables$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmccriseCanRepTables-grep_file"></a>}}
\if{latex}{\out{\hypertarget{method-UmccriseCanRepTables-grep_file}{}}}
\subsection{Method \code{grep_file()}}{
Returns file with given pattern from the cancer_report_tables directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmccriseCanRepTables$grep_file(pat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pat}}{File pattern to look for.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmccriseCanRepTables-read_chordtsv"></a>}}
\if{latex}{\out{\hypertarget{method-UmccriseCanRepTables-read_chordtsv}{}}}
\subsection{Method \code{read_chordtsv()}}{
Read \code{chord.tsv.gz} file output from umccrise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmccriseCanRepTables$read_chordtsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to \code{chord.tsv.gz} file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmccriseCanRepTables-read_hrdetecttsv"></a>}}
\if{latex}{\out{\hypertarget{method-UmccriseCanRepTables-read_hrdetecttsv}{}}}
\subsection{Method \code{read_hrdetecttsv()}}{
Read \code{hrdetect.tsv.gz} file output from umccrise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmccriseCanRepTables$read_hrdetecttsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to \code{hrdetect.tsv.gz} file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmccriseCanRepTables-read_sigs"></a>}}
\if{latex}{\out{\hypertarget{method-UmccriseCanRepTables-read_sigs}{}}}
\subsection{Method \code{read_sigs()}}{
Read \code{snv_20XX.tsv.gz} file output from umccrise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmccriseCanRepTables$read_sigs(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to \code{snv_20XX.tsv.gz} file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmccriseCanRepTables-read_qcsummarytsv"></a>}}
\if{latex}{\out{\hypertarget{method-UmccriseCanRepTables-read_qcsummarytsv}{}}}
\subsection{Method \code{read_qcsummarytsv()}}{
Read \code{qc_summary.tsv.gz} file output from umccrise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmccriseCanRepTables$read_qcsummarytsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character(1)})\cr
Path to \code{qc_summary.tsv.gz} file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmccriseCanRepTables-read"></a>}}
\if{latex}{\out{\hypertarget{method-UmccriseCanRepTables-read}{}}}
\subsection{Method \code{read()}}{
Reads contents of \code{cancer_report_tables} directory output by umccrise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmccriseCanRepTables$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of tibbles.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmccriseCanRepTables-write"></a>}}
\if{latex}{\out{\hypertarget{method-UmccriseCanRepTables-write}{}}}
\subsection{Method \code{write()}}{
Writes tidied contents of \code{cancer_report_tables} directory output by umccrise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmccriseCanRepTables$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmccriseCanRepTables-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UmccriseCanRepTables-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmccriseCanRepTables$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
