% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umccrise.R
\name{UmSigsSnvFile}
\alias{UmSigsSnvFile}
\title{UmSigsSnvFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{snv_20XX.tsv.gz} file with SNV signatures output from umccrise.
}
\examples{
\dontrun{
x <- "/path/to/snv_2015.tsv.gz"
d <- UmSigsSnvFile$new(x)
d_parsed <- d$read() # or read(d)
d$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "tsv")
}
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{UmSigsSnvFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UmSigsSnvFile-read}{\code{UmSigsSnvFile$read()}}
\item \href{#method-UmSigsSnvFile-write}{\code{UmSigsSnvFile$write()}}
\item \href{#method-UmSigsSnvFile-clone}{\code{UmSigsSnvFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmSigsSnvFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-UmSigsSnvFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{snv.tsv.gz} file output from umccrise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmSigsSnvFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmSigsSnvFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-UmSigsSnvFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{snv_20XX.tsv.gz} signature file output from umccrise.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmSigsSnvFile$write(d, out_dir, prefix, out_format = "tsv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UmSigsSnvFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UmSigsSnvFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UmSigsSnvFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
