% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wf.R
\name{Wf}
\alias{Wf}
\title{Workflow R6 Class}
\description{
Workflow is a base R6 class representing a bioinformatic
workflow run from a UMCCR workflow manager.
}
\examples{
p1 <- system.file("extdata/portaldb_workflow_top4.rds", package = "rportal") |>
  readRDS() |>
  dplyr::filter(type_name == "umccrise") |>
  dplyr::slice(1)
w <- Wf$new(
  prid = p1$portal_run_id, type = p1$type_name, start = p1$start, end = p1$end,
  status = p1$end_status, input = p1$input, output = p1$output
)
w
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prid}}{Portal run ID.}

\item{\code{type}}{Workflow type.}

\item{\code{start}}{Workflow start datetime.}

\item{\code{end}}{Workflow end datetime.}

\item{\code{status}}{Workflow end status.}

\item{\code{input}}{Workflow input JSON string.}

\item{\code{output}}{Workflow output JSON string.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf-new}{\code{Wf$new()}}
\item \href{#method-Wf-print}{\code{Wf$print()}}
\item \href{#method-Wf-clone}{\code{Wf$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf-new"></a>}}
\if{latex}{\out{\hypertarget{method-Wf-new}{}}}
\subsection{Method \code{new()}}{
Create a new Workflow object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf$new(
  prid = NULL,
  type = NULL,
  start = NULL,
  end = NULL,
  status = NULL,
  input = NULL,
  output = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prid}}{Portal run ID.}

\item{\code{type}}{Workflow type.}

\item{\code{start}}{Workflow start datetime.}

\item{\code{end}}{Workflow end datetime.}

\item{\code{status}}{Workflow end status.}

\item{\code{input}}{Workflow input JSON string.}

\item{\code{output}}{Workflow output JSON string.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf-print"></a>}}
\if{latex}{\out{\hypertarget{method-Wf-print}{}}}
\subsection{Method \code{print()}}{
Print details about the Workflow.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
