% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bclconvert.R
\name{Wf_bcl_convert}
\alias{Wf_bcl_convert}
\title{Wf bcl_convert R6 Class}
\description{
Contains methods for reading and processing files output from the UMCCR
\code{bcl_convert} workflow.
}
\examples{
\dontrun{
indir <- file.path()
sample_id <- "PTC_ctTSO240429"
library_id <- "L2400482"
d <- TsoCombinedVariantOutputFile$new(x)
d$read()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{indir}}{Input directory containing Reports per assay type
(e.g. /primary_data/240607_A01052_0209_BHLHFTDSXC/2024061140802544/).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf_bcl_convert-new}{\code{Wf_bcl_convert$new()}}
\item \href{#method-Wf_bcl_convert-print}{\code{Wf_bcl_convert$print()}}
\item \href{#method-Wf_bcl_convert-clone}{\code{Wf_bcl_convert$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_bcl_convert-new"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_bcl_convert-new}{}}}
\subsection{Method \code{new()}}{
Create a new Wf_bcl_convert object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_bcl_convert$new(indir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indir}}{Input directory containing Reports per assay type.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_bcl_convert-print"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_bcl_convert-print}{}}}
\subsection{Method \code{print()}}{
Print details about the Workflow
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_bcl_convert$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_bcl_convert-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_bcl_convert-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_bcl_convert$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
