% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso_dragen.R
\name{Wf_dragen}
\alias{Wf_dragen}
\title{Wf_dragen R6 Class}
\description{
Reads and writes tidy versions of files from the \code{dragen} workflow.
}
\examples{
\dontrun{

#---- Local ----#
prefix <- "L2401290"
p <- file.path(
  "~/s3/pipeline-prod-cache-503977275616-ap-southeast-2/byob-icav2/production",
  "analysis/cttsov2/20240915ff0295ed/Logs_Intermediates/DragenCaller",
  prefix
)
t1 <- Wf_dragen$new(path = p, prefix = prefix)
t1$list_files(max_files = 100)
t1$list_files_filter_relevant(max_files = 300)
d <- t1$download_files(max_files = 100, dryrun = F)
d_tidy <- t1$tidy_files(d)
d_write <- t1$write(
  d_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = prefix,
  format = "tsv"
)
}
}
\section{Super class}{
\code{\link[dracarys:Wf]{dracarys::Wf}} -> \code{Wf_dragen}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prefix}}{The LibraryID prefix of the sample (needed for path lookup).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf_dragen-new}{\code{Wf_dragen$new()}}
\item \href{#method-Wf_dragen-print}{\code{Wf_dragen$print()}}
\item \href{#method-Wf_dragen-read_replay}{\code{Wf_dragen$read_replay()}}
\item \href{#method-Wf_dragen-read_contigMeanCov}{\code{Wf_dragen$read_contigMeanCov()}}
\item \href{#method-Wf_dragen-read_coverageMetrics}{\code{Wf_dragen$read_coverageMetrics()}}
\item \href{#method-Wf_dragen-clone}{\code{Wf_dragen$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="download_files"><a href='../../dracarys/html/Wf.html#method-Wf-download_files'><code>dracarys::Wf$download_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files"><a href='../../dracarys/html/Wf.html#method-Wf-list_files'><code>dracarys::Wf$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files_filter_relevant"><a href='../../dracarys/html/Wf.html#method-Wf-list_files_filter_relevant'><code>dracarys::Wf$list_files_filter_relevant()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="tidy_files"><a href='../../dracarys/html/Wf.html#method-Wf-tidy_files'><code>dracarys::Wf$tidy_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="write"><a href='../../dracarys/html/Wf.html#method-Wf-write'><code>dracarys::Wf$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_dragen-new"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_dragen-new}{}}}
\subsection{Method \code{new()}}{
Create a new Wf_dragen object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_dragen$new(path = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to directory with raw workflow results (from S3 or
local filesystem).}

\item{\code{prefix}}{The LibraryID prefix of the sample (needed for path lookup).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_dragen-print"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_dragen-print}{}}}
\subsection{Method \code{print()}}{
Print details about the Workflow.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_dragen$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_dragen-read_replay"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_dragen-read_replay}{}}}
\subsection{Method \code{read_replay()}}{
Read \code{replay.json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_dragen$read_replay(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_dragen-read_contigMeanCov"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_dragen-read_contigMeanCov}{}}}
\subsection{Method \code{read_contigMeanCov()}}{
Read \code{contig_mean_cov.csv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_dragen$read_contigMeanCov(x, keep_alt = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}

\item{\code{keep_alt}}{Keep ALT contigs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_dragen-read_coverageMetrics"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_dragen-read_coverageMetrics}{}}}
\subsection{Method \code{read_coverageMetrics()}}{
Read \code{dragen.tsv.gz} cancer report hrd file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_dragen$read_coverageMetrics(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_dragen-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_dragen-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_dragen$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
