% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sash.R
\name{Wf_sash}
\alias{Wf_sash}
\title{Wf_sash R6 Class}
\description{
Reads and writes tidy versions of files from the \code{sash} workflow
}
\examples{
\dontrun{

#---- Local ----#
p1 <- "~/s3/org.umccr.data.oncoanalyser/analysis_data/SBJ05571/sash"
p2 <- "202408270b93455e/L2401308_L2401307"
p <- normalizePath(file.path(p1, p2))
SubjectID <- "SBJ05571"
SampleID_tumor <- "MDX240307"
prefix <- glue("{SubjectID}__{SampleID_tumor}")
s1 <- Wf_sash$new(path = p, SubjectID = SubjectID, SampleID_tumor = SampleID_tumor)
s1$list_files(max_files = 20)
s1$list_files_filter_relevant(max_files = 300)
d <- s1$download_files(max_files = 1000, dryrun = F)
d_tidy <- s1$tidy_files(d)
d_write <- s1$write(
  d_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = glue("{SubjectID}_{SampleID_tumor}"),
  format = "tsv"
)

#---- S3 ----#
p1 <- "s3://org.umccr.data.oncoanalyser/analysis_data/SBJ05571/sash"
p2 <- "202408270b93455e/L2401308_L2401307"
p <- file.path(p1, p2)
SubjectID <- "SBJ05571"
SampleID_tumor <- "MDX240307"
prefix <- glue("{SubjectID}__{SampleID_tumor}")
s1 <- Wf_sash$new(path = p, SubjectID = SubjectID, SampleID_tumor = SampleID_tumor)
s1$list_files(max_files = 20)
s1$list_files_filter_relevant()
outdir <- sub("s3:/", "~/s3", p)
d <- s1$download_files(outdir = outdir, max_files = 1000, dryrun = F)
d_tidy <- s1$tidy_files(d)
d_write <- s1$write(
  d_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = glue("{SubjectID}__{SampleID_tumor}"),
  format = "tsv"
)
}

}
\section{Super class}{
\code{\link[dracarys:Wf]{dracarys::Wf}} -> \code{Wf_sash}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{SubjectID}}{The SubjectID of the sample (needed for path lookup).}

\item{\code{SampleID_tumor}}{The SampleID of the tumor sample (needed for path lookup).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf_sash-new}{\code{Wf_sash$new()}}
\item \href{#method-Wf_sash-print}{\code{Wf_sash$print()}}
\item \href{#method-Wf_sash-list_files_filter_relevant}{\code{Wf_sash$list_files_filter_relevant()}}
\item \href{#method-Wf_sash-download_files}{\code{Wf_sash$download_files()}}
\item \href{#method-Wf_sash-read_pcgrjson}{\code{Wf_sash$read_pcgrjson()}}
\item \href{#method-Wf_sash-read_hrddragentsv}{\code{Wf_sash$read_hrddragentsv()}}
\item \href{#method-Wf_sash-read_hrdchordtsv}{\code{Wf_sash$read_hrdchordtsv()}}
\item \href{#method-Wf_sash-read_hrdetecttsv}{\code{Wf_sash$read_hrdetecttsv()}}
\item \href{#method-Wf_sash-read_sigstsv}{\code{Wf_sash$read_sigstsv()}}
\item \href{#method-Wf_sash-read_sigssnv2015tsv}{\code{Wf_sash$read_sigssnv2015tsv()}}
\item \href{#method-Wf_sash-read_sigssnv2020tsv}{\code{Wf_sash$read_sigssnv2020tsv()}}
\item \href{#method-Wf_sash-read_sigsdbstsv}{\code{Wf_sash$read_sigsdbstsv()}}
\item \href{#method-Wf_sash-read_sigsindeltsv}{\code{Wf_sash$read_sigsindeltsv()}}
\item \href{#method-Wf_sash-read_qcsummarytsv}{\code{Wf_sash$read_qcsummarytsv()}}
\item \href{#method-Wf_sash-clone}{\code{Wf_sash$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files"><a href='../../dracarys/html/Wf.html#method-Wf-list_files'><code>dracarys::Wf$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="tidy_files"><a href='../../dracarys/html/Wf.html#method-Wf-tidy_files'><code>dracarys::Wf$tidy_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="write"><a href='../../dracarys/html/Wf.html#method-Wf-write'><code>dracarys::Wf$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-new"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-new}{}}}
\subsection{Method \code{new()}}{
Create a new Wf_sash object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$new(path = NULL, SubjectID = NULL, SampleID_tumor = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to directory with raw workflow results (from GDS, S3, or
local filesystem).}

\item{\code{SubjectID}}{The SubjectID of the sample (needed for path lookup).}

\item{\code{SampleID_tumor}}{The SampleID of the tumor sample (needed for path lookup).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-print"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-print}{}}}
\subsection{Method \code{print()}}{
Print details about the Workflow.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-list_files_filter_relevant"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-list_files_filter_relevant}{}}}
\subsection{Method \code{list_files_filter_relevant()}}{
List dracarys files under given path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$list_files_filter_relevant(
  max_files = 1000,
  ica_token = Sys.getenv("ICA_ACCESS_TOKEN"),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{max_files}}{Max number of files to list (for gds/s3 only).}

\item{\code{ica_token}}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}

\item{\code{...}}{Passed on to the \code{gds_list_files_filter_relevant} or
the \code{s3_list_files_filter_relevant} function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-download_files"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-download_files}{}}}
\subsection{Method \code{download_files()}}{
Download files from GDS/S3 to local filesystem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$download_files(
  outdir,
  ica_token = Sys.getenv("ICA_ACCESS_TOKEN"),
  max_files = 1000,
  dryrun = FALSE,
  recursive = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outdir}}{Path to output directory.}

\item{\code{ica_token}}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}

\item{\code{max_files}}{Max number of files to list.}

\item{\code{dryrun}}{If TRUE, just list the files that will be downloaded (don't
download them).}

\item{\code{recursive}}{Should files be returned recursively \emph{in and under} the specified
GDS directory, or \emph{only directly in} the specified GDS directory (def: TRUE via ICA API).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_pcgrjson"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_pcgrjson}{}}}
\subsection{Method \code{read_pcgrjson()}}{
Read \code{pcgr.json.gz} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_pcgrjson(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_hrddragentsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_hrddragentsv}{}}}
\subsection{Method \code{read_hrddragentsv()}}{
Read \code{dragen.tsv.gz} cancer report hrd file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_hrddragentsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_hrdchordtsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_hrdchordtsv}{}}}
\subsection{Method \code{read_hrdchordtsv()}}{
Read \code{chord.tsv.gz} cancer report hrd file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_hrdchordtsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_hrdetecttsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_hrdetecttsv}{}}}
\subsection{Method \code{read_hrdetecttsv()}}{
Read \code{hrdetect.tsv.gz} cancer report hrd file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_hrdetecttsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_sigstsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_sigstsv}{}}}
\subsection{Method \code{read_sigstsv()}}{
Read signature cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_sigstsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_sigssnv2015tsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_sigssnv2015tsv}{}}}
\subsection{Method \code{read_sigssnv2015tsv()}}{
Read \code{snv_2015.tsv.gz} sigs cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_sigssnv2015tsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_sigssnv2020tsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_sigssnv2020tsv}{}}}
\subsection{Method \code{read_sigssnv2020tsv()}}{
Read \code{snv_2020.tsv.gz} sigs cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_sigssnv2020tsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_sigsdbstsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_sigsdbstsv}{}}}
\subsection{Method \code{read_sigsdbstsv()}}{
Read \code{dbs.tsv.gz} sigs cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_sigsdbstsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_sigsindeltsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_sigsindeltsv}{}}}
\subsection{Method \code{read_sigsindeltsv()}}{
Read \code{indel.tsv.gz} sigs cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_sigsindeltsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_qcsummarytsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_qcsummarytsv}{}}}
\subsection{Method \code{read_qcsummarytsv()}}{
Read \code{qc_summary.tsv.gz} cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_qcsummarytsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
