% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{Wf_tso_ctdna_tumor_only}
\alias{Wf_tso_ctdna_tumor_only}
\title{tso_ctdna_tumor_only Wf R6 Class}
\description{
Contains methods for reading and processing files output from the UMCCR
\code{tso_ctdna_tumor_only} workflow.
}
\examples{
\dontrun{
x <- file.path(
  "~/icav1/g/production/analysis_data/SBJ00596/tso_ctdna_tumor_only",
  "2024050555972acf/L2400482/Results/PTC_ctTSO240429_L2400482/dracarys_gds_sync"
  )
sample_id <- "PTC_ctTSO240429"
library_id <- "L2400482"
d <- TsoCombinedVariantOutputFile$new(x)
d$read()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sid}}{SampleID.}

\item{\code{lid}}{LibraryID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf_tso_ctdna_tumor_only-clone}{\code{Wf_tso_ctdna_tumor_only$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
