% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsov2.R
\name{Wf_tso_ctdna_tumor_only_v2}
\alias{Wf_tso_ctdna_tumor_only_v2}
\title{Wf_tso_ctdna_tumor_only_v2 R6 Class}
\description{
Reads and writes tidy versions of files from the \code{tso_ctdna_tumor_only_v2} workflow.
}
\examples{
\dontrun{

#---- Local ----#
p <- file.path(
  "~/s3/pipeline-prod-cache-503977275616-ap-southeast-2/byob-icav2/production",
  "analysis/cttsov2/20240915ff0295ed"
)
prefix <- "L2401290"
t1 <- Wf_tso_ctdna_tumor_only_v2$new(path = p, prefix = prefix)
t1$list_files(max_files = 100)
t1$dragenObj$list_files(max_files = 100)
t1$list_files_filter_relevant(max_files = 300)
t1$dragenObj$list_files_filter_relevant(max_files = 300)
d1 <- t1$download_files(max_files = 100, dryrun = F)
d2 <- t1$dragenObj$download_files(max_files = 100, dryrun = F)
d1_tidy <- t1$tidy_files(d1)
d2_tidy <- t1$dragenObj$tidy_files(d2)
d_write1 <- t1$write(
  d1_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = prefix,
  format = "tsv"
)
d_write2 <- t1$dragenObj$write(
  d2_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = prefix,
  format = "tsv"
)

#---- S3 ----#
p <- file.path(
  "s3://pipeline-prod-cache-503977275616-ap-southeast-2/byob-icav2/production",
  "analysis/cttsov2/20240915ff0295ed"
)
prefix <- "L2401290"
outdir <- sub("s3:/", "~/s3", p)
t2 <- Wf_tso_ctdna_tumor_only_v2$new(path = p, prefix = prefix)
t2$list_files(max_files = 500)
t2$list_files_filter_relevant(max_files = 500)
d <- t2$download_files(
  outdir = outdir,
  max_files = 500,
  dryrun = FALSE
)
d_tidy <- t2$tidy_files(d)
d_write <- t2$write(
  d_tidy,
  outdir = file.path(outdir, "dracarys_tidy"),
  prefix = prefix,
  format = "tsv"
)
}
}
\section{Super class}{
\code{\link[dracarys:Wf]{dracarys::Wf}} -> \code{Wf_tso_ctdna_tumor_only_v2}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{prefix}}{The LibraryID prefix of the tumor sample (needed for path lookup).}

\item{\code{dragenObj}}{dragen object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-new}{\code{Wf_tso_ctdna_tumor_only_v2$new()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-print}{\code{Wf_tso_ctdna_tumor_only_v2$print()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-read_sar}{\code{Wf_tso_ctdna_tumor_only_v2$read_sar()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-read_tmbt}{\code{Wf_tso_ctdna_tumor_only_v2$read_tmbt()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-read_cvo}{\code{Wf_tso_ctdna_tumor_only_v2$read_cvo()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-read_cnv}{\code{Wf_tso_ctdna_tumor_only_v2$read_cnv()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-read_cvgrepe}{\code{Wf_tso_ctdna_tumor_only_v2$read_cvgrepe()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-read_cvgrepg}{\code{Wf_tso_ctdna_tumor_only_v2$read_cvgrepg()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-read_hardfilt}{\code{Wf_tso_ctdna_tumor_only_v2$read_hardfilt()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-read_msi}{\code{Wf_tso_ctdna_tumor_only_v2$read_msi()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-read_fus}{\code{Wf_tso_ctdna_tumor_only_v2$read_fus()}}
\item \href{#method-Wf_tso_ctdna_tumor_only_v2-clone}{\code{Wf_tso_ctdna_tumor_only_v2$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="DOWNLOAD_ONLY"><a href='../../dracarys/html/Wf.html#method-Wf-DOWNLOAD_ONLY'><code>dracarys::Wf$DOWNLOAD_ONLY()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="download_files"><a href='../../dracarys/html/Wf.html#method-Wf-download_files'><code>dracarys::Wf$download_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files"><a href='../../dracarys/html/Wf.html#method-Wf-list_files'><code>dracarys::Wf$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files_filter_relevant"><a href='../../dracarys/html/Wf.html#method-Wf-list_files_filter_relevant'><code>dracarys::Wf$list_files_filter_relevant()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="tidy_files"><a href='../../dracarys/html/Wf.html#method-Wf-tidy_files'><code>dracarys::Wf$tidy_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="write"><a href='../../dracarys/html/Wf.html#method-Wf-write'><code>dracarys::Wf$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-new"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-new}{}}}
\subsection{Method \code{new()}}{
Create a new Wf_tso_ctdna_tumor_only_v2 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$new(path = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to directory with raw workflow results (from S3 or
local filesystem).}

\item{\code{prefix}}{The LibraryID prefix of the tumor sample (needed for path lookup).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-print"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-print}{}}}
\subsection{Method \code{print()}}{
Print details about the Workflow.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-read_sar"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-read_sar}{}}}
\subsection{Method \code{read_sar()}}{
Read \code{SampleAnalysisResults.json.gz} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$read_sar(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-read_tmbt"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-read_tmbt}{}}}
\subsection{Method \code{read_tmbt()}}{
Read \code{TMB_Trace.tsv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$read_tmbt(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-read_cvo"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-read_cvo}{}}}
\subsection{Method \code{read_cvo()}}{
Read \code{CombinedVariantOutput.tsv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$read_cvo(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-read_cnv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-read_cnv}{}}}
\subsection{Method \code{read_cnv()}}{
Read \code{cnv.vcf} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$read_cnv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-read_cvgrepe"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-read_cvgrepe}{}}}
\subsection{Method \code{read_cvgrepe()}}{
Read \code{exon_cov_report.tsv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$read_cvgrepe(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-read_cvgrepg"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-read_cvgrepg}{}}}
\subsection{Method \code{read_cvgrepg()}}{
Read \code{gene_cov_report.tsv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$read_cvgrepg(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-read_hardfilt"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-read_hardfilt}{}}}
\subsection{Method \code{read_hardfilt()}}{
Read \code{hard-filtered.vcf} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$read_hardfilt(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-read_msi"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-read_msi}{}}}
\subsection{Method \code{read_msi()}}{
Read \code{microsat_output.json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$read_msi(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-read_fus"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-read_fus}{}}}
\subsection{Method \code{read_fus()}}{
Read \code{Fusions.csv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$read_fus(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_tso_ctdna_tumor_only_v2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_tso_ctdna_tumor_only_v2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_tso_ctdna_tumor_only_v2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
