% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umccrise.R
\name{Wf_umccrise}
\alias{Wf_umccrise}
\title{Wf_umccrise R6 Class}
\description{
Reads and writes tidy versions of files from the \code{umccrise} workflow
}
\examples{
\dontrun{

#---- LOCAL ----#
SubjectID <- "SBJ03043"
SampleID_tumor <- "PRJ230004"
prefix <- glue("{SubjectID}__{SampleID_tumor}")
p1_local <- "~/icav1/g/production/analysis_data"
p <- file.path(p1_local, "SBJ03043/umccrise/20240830ec648f40/L2300064__L2300063")
um1 <- Wf_umccrise$new(path = p, SubjectID = SubjectID, SampleID_tumor = SampleID_tumor)
um1$list_files(max_files = 10)
um1$list_files_filter_relevant()
d <- um1$download_files(max_files = 1000, dryrun = F)
d_tidy <- um1$tidy_files(d)
d_write <- um1$write(
  d_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = glue("{SubjectID}__{SampleID_tumor}"),
  format = "tsv"
)

#---- GDS ----#
SubjectID <- "SBJ03043"
SampleID_tumor <- "PRJ230004"
prefix <- glue("{SubjectID}__{SampleID_tumor}")
p1_gds <- "gds://production/analysis_data"
p <- file.path(p1_gds, "SBJ03043/umccrise/20240830ec648f40/L2300064__L2300063")
outdir <- file.path(sub("gds:/", "~/icav1/g", p))
token <- Sys.getenv("ICA_ACCESS_TOKEN")
um2 <- Wf_umccrise$new(path = p, SubjectID = SubjectID, SampleID_tumor = SampleID_tumor)
um2$list_files(max_files = 8)
um2$list_files_filter_relevant(ica_token = token, max_files = 500)
d <- um2$download_files(
  outdir = outdir, ica_token = token,
  max_files = 1000, dryrun = F
)
d_tidy <- um2$tidy_files(d)
d_write <- um2$write(
  d_tidy,
  outdir = file.path(outdir, "dracarys_tidy"),
  prefix = glue("{SubjectID}__{SampleID_tumor}"),
  format = "tsv"
)
}

}
\section{Super class}{
\code{\link[dracarys:Wf]{dracarys::Wf}} -> \code{Wf_umccrise}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{SubjectID}}{The SubjectID of the sample (needed for path lookup).}

\item{\code{SampleID_tumor}}{The SampleID of the tumor sample (needed for path lookup).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf_umccrise-new}{\code{Wf_umccrise$new()}}
\item \href{#method-Wf_umccrise-print}{\code{Wf_umccrise$print()}}
\item \href{#method-Wf_umccrise-read_pcgr_json}{\code{Wf_umccrise$read_pcgr_json()}}
\item \href{#method-Wf_umccrise-read_hrd_chord}{\code{Wf_umccrise$read_hrd_chord()}}
\item \href{#method-Wf_umccrise-read_hrd_hrdetect}{\code{Wf_umccrise$read_hrd_hrdetect()}}
\item \href{#method-Wf_umccrise-read_sigstsv}{\code{Wf_umccrise$read_sigstsv()}}
\item \href{#method-Wf_umccrise-read_sigs_snv2015}{\code{Wf_umccrise$read_sigs_snv2015()}}
\item \href{#method-Wf_umccrise-read_sigs_snv2020}{\code{Wf_umccrise$read_sigs_snv2020()}}
\item \href{#method-Wf_umccrise-read_sigs_dbs}{\code{Wf_umccrise$read_sigs_dbs()}}
\item \href{#method-Wf_umccrise-read_sigs_indel}{\code{Wf_umccrise$read_sigs_indel()}}
\item \href{#method-Wf_umccrise-read_qcsum}{\code{Wf_umccrise$read_qcsum()}}
\item \href{#method-Wf_umccrise-read_conpairmultiqc}{\code{Wf_umccrise$read_conpairmultiqc()}}
\item \href{#method-Wf_umccrise-clone}{\code{Wf_umccrise$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="download_files"><a href='../../dracarys/html/Wf.html#method-Wf-download_files'><code>dracarys::Wf$download_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files"><a href='../../dracarys/html/Wf.html#method-Wf-list_files'><code>dracarys::Wf$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files_filter_relevant"><a href='../../dracarys/html/Wf.html#method-Wf-list_files_filter_relevant'><code>dracarys::Wf$list_files_filter_relevant()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="tidy_files"><a href='../../dracarys/html/Wf.html#method-Wf-tidy_files'><code>dracarys::Wf$tidy_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="write"><a href='../../dracarys/html/Wf.html#method-Wf-write'><code>dracarys::Wf$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-new"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-new}{}}}
\subsection{Method \code{new()}}{
Create a new Wf_umccrise object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$new(path = NULL, SubjectID = NULL, SampleID_tumor = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to directory with raw workflow results (from GDS, S3, or
local filesystem).}

\item{\code{SubjectID}}{The SubjectID of the sample (needed for path lookup).}

\item{\code{SampleID_tumor}}{The SampleID of the tumor sample (needed for path lookup).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-print"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-print}{}}}
\subsection{Method \code{print()}}{
Print details about the Workflow.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_pcgr_json"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_pcgr_json}{}}}
\subsection{Method \code{read_pcgr_json()}}{
Read \code{pcgr.json.gz} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_pcgr_json(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_hrd_chord"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_hrd_chord}{}}}
\subsection{Method \code{read_hrd_chord()}}{
Read \code{chord.tsv.gz} cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_hrd_chord(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_hrd_hrdetect"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_hrd_hrdetect}{}}}
\subsection{Method \code{read_hrd_hrdetect()}}{
Read \code{hrdetect.tsv.gz} cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_hrd_hrdetect(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_sigstsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_sigstsv}{}}}
\subsection{Method \code{read_sigstsv()}}{
Read signature cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_sigstsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_sigs_snv2015"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_sigs_snv2015}{}}}
\subsection{Method \code{read_sigs_snv2015()}}{
Read \code{snv_2015.tsv.gz} sigs cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_sigs_snv2015(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_sigs_snv2020"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_sigs_snv2020}{}}}
\subsection{Method \code{read_sigs_snv2020()}}{
Read \code{snv_2020.tsv.gz} sigs cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_sigs_snv2020(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_sigs_dbs"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_sigs_dbs}{}}}
\subsection{Method \code{read_sigs_dbs()}}{
Read \code{dbs.tsv.gz} sigs cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_sigs_dbs(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_sigs_indel"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_sigs_indel}{}}}
\subsection{Method \code{read_sigs_indel()}}{
Read \code{indel.tsv.gz} sigs cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_sigs_indel(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_qcsum"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_qcsum}{}}}
\subsection{Method \code{read_qcsum()}}{
Read \code{qc_summary.tsv.gz} cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_qcsum(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_conpairmultiqc"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_conpairmultiqc}{}}}
\subsection{Method \code{read_conpairmultiqc()}}{
Read multiqc_conpair.txt file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_conpairmultiqc(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
