% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umccrise.R
\name{Wf_umccrise_download_tidy_write}
\alias{Wf_umccrise_download_tidy_write}
\title{umccrise Download Tidy and Write}
\usage{
Wf_umccrise_download_tidy_write(
  path,
  SubjectID,
  SampleID_tumor,
  SampleID_normal,
  outdir,
  format = "rds",
  max_files = 1000,
  ica_token = Sys.getenv("ICA_ACCESS_TOKEN"),
  dryrun = FALSE
)
}
\arguments{
\item{path}{Path to directory with raw workflow results (from GDS, S3, or
local filesystem).}

\item{SubjectID}{The SubjectID of the sample.}

\item{SampleID_tumor}{The SampleID of the tumor sample.}

\item{SampleID_normal}{The SampleID of the normal sample.}

\item{outdir}{Path to output directory.}

\item{format}{Format of output files.}

\item{max_files}{Max number of files to list.}

\item{ica_token}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}

\item{dryrun}{If TRUE, just list the files that will be downloaded (don't
download them).}
}
\value{
Tibble of tidy data as list-cols.
}
\description{
Downloads files from the \code{umccrise} workflow and writes them in a tidy format.
}
\examples{
\dontrun{
SubjectID <- "SBJ04662"
SampleID_tumor <- "PRJ240647"
SampleID_normal <- "PRJ240646"
p1_gds <- "gds://production/analysis_data"
p <- file.path(p1_gds, "SBJ04662/umccrise/20240302e66750fe/L2400240__L2400239")
outdir <- file.path(sub("gds:/", "~/icav1/g", p))
token <- Sys.getenv("ICA_ACCESS_TOKEN")
d <- Wf_umccrise_download_tidy_write(
  path = p, SubjectID = SubjectID,
  SampleID_tumor = SampleID_tumor, SampleID_normal = SampleID_normal,
  outdir = outdir,
  dryrun = F
)
}
}
