% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_dracarys_list_of_tbls}
\alias{write_dracarys_list_of_tbls}
\title{Write List of Tidy Tibbles}
\usage{
write_dracarys_list_of_tbls(
  list_of_tbls,
  out_dir = NULL,
  prefix = NULL,
  out_format = "tsv",
  drid = NULL
)
}
\arguments{
\item{list_of_tbls}{List of tidy tibbles.}

\item{out_dir}{Output directory.}

\item{prefix}{Prefix of output file(s).}

\item{out_format}{Format of output file(s).}

\item{drid}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\value{
Tibble with nested objects that have been written to the output directory.
}
\description{
Write List of Tidy Tibbles
}
