% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf.R
\name{bcftools_parse_vcf}
\alias{bcftools_parse_vcf}
\title{Parse VCF with bcftools}
\usage{
bcftools_parse_vcf(vcf, only_pass = TRUE)
}
\arguments{
\item{vcf}{VCF with one or more samples.}

\item{only_pass}{Keep PASS variants only (TRUE by default)?}
}
\value{
A tibble with all the main, FORMAT, and INFO fields detected in
the VCF header as columns.
}
\description{
Parse VCF with bcftools.
Uses bcftools under the hood to do the heavy lifting with field splitting,
then converts the parsed character vector to a tibble.
}
