% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContigMeanCovFile.R
\name{ContigMeanCovFile}
\alias{ContigMeanCovFile}
\title{ContigMeanCovFile R6 Class}
\description{
Contains methods for reading and displaying contents of
the \verb{wgs_contig_mean_cov_<phenotype>.csv} file output from DRAGEN.
This file contains the estimated coverage for all contigs, and an autosomal
estimated coverage.
}
\examples{
x1 <- system.file("extdata/wgs/SEQC-II.wgs_contig_mean_cov_normal.csv.gz", package = "dracarys")
x2 <- system.file("extdata/wgs/SEQC-II.wgs_contig_mean_cov_tumor.csv.gz", package = "dracarys")
cc1 <- ContigMeanCovFile$new(x1)
cc2 <- ContigMeanCovFile$new(x2)
read(cc1)
read(cc2, keep_alt = TRUE)
plot(cc1)

}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{ContigMeanCovFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContigMeanCovFile-read}{\code{ContigMeanCovFile$read()}}
\item \href{#method-ContigMeanCovFile-plot}{\code{ContigMeanCovFile$plot()}}
\item \href{#method-ContigMeanCovFile-clone}{\code{ContigMeanCovFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContigMeanCovFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-ContigMeanCovFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \verb{wgs_contig_mean_cov_<phenotype>.csv} file output from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContigMeanCovFile$read(keep_alt = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keep_alt}}{Keep the ALT + Mito chromosomes?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble with the following columns:
\itemize{
\item label: file label.
\item chrom: contig name.
\item n_bases: number of bases aligned to contig (excludes bases from
duplicate marked reads, reads with MAPQ=0, and clipped bases).
\item coverage: col2 / contig length
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContigMeanCovFile-plot"></a>}}
\if{latex}{\out{\hypertarget{method-ContigMeanCovFile-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the \verb{wgs_contig_mean_cov_<phenotype>.csv} files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContigMeanCovFile$plot(top_alt_n = 15)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{top_alt_n}}{Number of top covered alt contigs to plot per phenotype.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object with chromosomes on X axis, and coverage on Y axis.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContigMeanCovFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ContigMeanCovFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContigMeanCovFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
