% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{CoverageMetricsFile}
\alias{CoverageMetricsFile}
\title{CoverageMetricsFile R6 Class}
\description{
Contains methods for reading and displaying contents of
the \verb{wgs_coverage_metrics_<phenotype>.csv} file output from DRAGEN.
This file contains read depth of coverage metrics.
}
\examples{
x1 <- system.file("extdata/wgs/SEQC-II.wgs_coverage_metrics_normal.csv.gz", package = "dracarys")
x2 <- system.file("extdata/wgs/SEQC-II.wgs_coverage_metrics_tumor.csv.gz", package = "dracarys")
cm1 <- CoverageMetricsFile$new(x1)
cm2 <- CoverageMetricsFile$new(x2)
d1 <- read(cm1)
d2 <- read(cm2)
cm1$write(d1, out_dir = tempdir(), prefix = "seqc_n", out_format = "tsv")
cm2$write(d2, out_dir = tempdir(), prefix = "seqc_t", out_format = "tsv")

}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{CoverageMetricsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CoverageMetricsFile-read}{\code{CoverageMetricsFile$read()}}
\item \href{#method-CoverageMetricsFile-write}{\code{CoverageMetricsFile$write()}}
\item \href{#method-CoverageMetricsFile-clone}{\code{CoverageMetricsFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoverageMetricsFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-CoverageMetricsFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \verb{wgs_coverage_metrics_<phenotype>.csv} file output from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverageMetricsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item label: file label.
\item var: variable name.
\item var_abbrev: variable abbreviation.
\item pct: percentage value.
\item count: count value.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoverageMetricsFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-CoverageMetricsFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \verb{wgs_coverage_metrics_<phenotype>.csv} file output
from DRAGEN
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverageMetricsFile$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CoverageMetricsFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CoverageMetricsFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CoverageMetricsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
