% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_test}
\alias{db_test}
\title{Test DB}
\usage{
db_test(
  path,
  prefix,
  outdir,
  prid,
  max_files,
  dbname = "nemo",
  dbuser = "orcabus"
)
}
\description{
Test DB
}
\examples{
\dontrun{
path <- file.path(
  "~/s3/pipeline-prod-cache-503977275616-ap-southeast-2/byob-icav2/production",
  "analysis/wgts-qc/20241123ffa837c4/L2401621_dragen_alignment"
)
path <- file.path(
  "~/s3/pipeline-prod-cache-503977275616-ap-southeast-2/byob-icav2/production",
  "analysis/cttsov2/20250308f448d4e0"
)
prefix <- "L2500183" # tsov2
prefix <- "L2401621" # alignqc
outdir <- path
max_files <- 1000
prid <- "20250308f448d4e0"
prid <- "efgh5678"
dbname <- "nemo"
dbuser <- "orcabus"
d <- db_test(
  path = path, prefix = prefix, outdir = outdir,
  prid = prid, max_files = max_files,
  dbname = dbname, dbuser = dbuser
)
}

}
