% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{dr_gds_download}
\alias{dr_gds_download}
\title{dracarys GDS Download}
\usage{
dr_gds_download(
  gdsdir,
  outdir,
  token,
  page_size = 100,
  pattern = NULL,
  dryrun = FALSE,
  regexes = DR_FILE_REGEX
)
}
\arguments{
\item{gdsdir}{Full path to GDS directory.}

\item{outdir}{Path to output directory.}

\item{token}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}

\item{page_size}{Page size (def: 100).}

\item{pattern}{Pattern to further filter the returned file type tibble.}

\item{dryrun}{If TRUE, just list the files that will be downloaded (don't
download them).}
}
\description{
Download only GDS files that can be processed by dracarys.
}
