% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{dracarys_tidy_multiqc}
\alias{dracarys_tidy_multiqc}
\title{Dracarys Tidy MultiQC}
\usage{
dracarys_tidy_multiqc(json, prefix, outdir)
}
\arguments{
\item{json}{Path to \code{multiqc_data.json}.}

\item{prefix}{Prefix for output files.}

\item{outdir}{Path to output results.}
}
\value{
Generates TSV and Parquet representations of the input
MultiQC JSON file.
}
\description{
Generate tidier representations of MultiQC JSON output
}
