% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File.R
\name{File}
\alias{File}
\title{File R6 Class}
\description{
File is a base R6 class representing a TSV/CSV/JSON output from
a DRAGEN workflow.

A File has a path, a basename, a type, and a default read method for its type.
}
\examples{
F1 <- File$new(readr::readr_example("mtcars.csv"))
(parsed_f1 <- F1$read(col_types = readr::cols("double")))
(bname_f1 <- F1$bname())
(F2 <- File$new("https://stratus-gds-aps2/foo/bar/baz.csv?bla"))

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Name or full path of the file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-File-new}{\code{File$new()}}
\item \href{#method-File-bname}{\code{File$bname()}}
\item \href{#method-File-type}{\code{File$type()}}
\item \href{#method-File-print}{\code{File$print()}}
\item \href{#method-File-read}{\code{File$read()}}
\item \href{#method-File-clone}{\code{File$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-new"></a>}}
\if{latex}{\out{\hypertarget{method-File-new}{}}}
\subsection{Method \code{new()}}{
Create a new File object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$new(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Name or full path of the file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-bname"></a>}}
\if{latex}{\out{\hypertarget{method-File-bname}{}}}
\subsection{Method \code{bname()}}{
Basename of the file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$bname()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Basename of the file as a character vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-type"></a>}}
\if{latex}{\out{\hypertarget{method-File-type}{}}}
\subsection{Method \code{type()}}{
Get the type of file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
String describing the type of file (CSV, TSV, JSON or OTHER).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-print"></a>}}
\if{latex}{\out{\hypertarget{method-File-print}{}}}
\subsection{Method \code{print()}}{
Print details about the File.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-read"></a>}}
\if{latex}{\out{\hypertarget{method-File-read}{}}}
\subsection{Method \code{read()}}{
Read the file based on its type.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$read(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed on to appropriate read_* function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-File-clone"></a>}}
\if{latex}{\out{\hypertarget{method-File-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{File$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
