% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FineHistFile.R
\name{FineHistFile}
\alias{FineHistFile}
\title{FineHistFile R6 Class}
\description{
Contains methods for reading and displaying contents of
the \verb{wgs_fine_hist_<phenotype>.csv} file output from DRAGEN.
This file contains two columns: Depth and Overall.
The value in the Depth column ranges from 0 to 1000+ and the Overall
column indicates the number of loci covered at the corresponding depth.
}
\examples{
x1 <- system.file("extdata/wgs/SEQC-II.wgs_fine_hist_normal.csv.gz", package = "dracarys")
x2 <- system.file("extdata/wgs/SEQC-II.wgs_fine_hist_tumor.csv.gz", package = "dracarys")
ch1 <- FineHistFile$new(x1)
ch2 <- FineHistFile$new(x2)
read(ch1)
read(ch2)
plot(ch1)
plot(ch2)

}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{FineHistFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-read}{\code{FineHistFile$read()}}
\item \href{#method-plot}{\code{FineHistFile$plot()}}
\item \href{#method-clone}{\code{FineHistFile$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname">}\href{../../dracarys/html/File.html#method-bname}{\code{dracarys::File$bname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize">}\href{../../dracarys/html/File.html#method-initialize}{\code{dracarys::File$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print">}\href{../../dracarys/html/File.html#method-print}{\code{dracarys::File$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type">}\href{../../dracarys/html/File.html#method-type}{\code{dracarys::File$type()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Reads the \verb{wgs_fine_hist_<phenotype>.csv} file output from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineHistFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with three columns:
\itemize{
\item label
\item depth
\item number of loci with given depth
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the \verb{wgs_fine_hist_<phenotype>.csv} files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineHistFile$plot(x_lim = c(0, 300))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_lim}}{X axis range to plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 object with depth of coverage on X axis,
and number of loci with that depth on Y axis.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FineHistFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
