% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentLengthHistFile.R
\name{FragmentLengthHistFile}
\alias{FragmentLengthHistFile}
\title{FragmentLengthHistFile R6 Class}
\description{
Contains methods for reading and plotting contents of
the \code{fragment_length_hist.csv} file output from DRAGEN.
}
\examples{
x <- system.file("extdata/wgs/SEQC-II.fragment_length_hist.csv.gz", package = "dracarys")
fl <- FragmentLengthHistFile$new(x)
fl$read() # or read(fl)
fl$plot() # or plot(fl)
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{FragmentLengthHistFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-read}{\code{FragmentLengthHistFile$read()}}
\item \href{#method-plot}{\code{FragmentLengthHistFile$plot()}}
\item \href{#method-clone}{\code{FragmentLengthHistFile$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname">}\href{../../dracarys/html/File.html#method-bname}{\code{dracarys::File$bname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize">}\href{../../dracarys/html/File.html#method-initialize}{\code{dracarys::File$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print">}\href{../../dracarys/html/File.html#method-print}{\code{dracarys::File$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type">}\href{../../dracarys/html/File.html#method-type}{\code{dracarys::File$type()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{fragment_length_hist.csv} file, which contains the
fragment length distribution for each sample.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FragmentLengthHistFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble with the following columns:
\itemize{
\item fragmentLength: estimated fragment length
\item count: number of reads with estimated fragment length
\item sample: name of sample
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the fragment length distributions as given in the
\code{fragment_length_hist.csv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FragmentLengthHistFile$plot(min_count = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_count}}{Minimum read count to be plotted (Default: 10).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 plot containing fragment lengths on X axis and read counts
on Y axis for each sample.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FragmentLengthHistFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
