% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{gds_files_list}
\alias{gds_files_list}
\title{GDS Files List}
\usage{
gds_files_list(gdsdir, token = Sys.getenv("ICA_ACCESS_TOKEN"))
}
\arguments{
\item{gdsdir}{Full path to GDS directory.}

\item{token}{ICA access token (by default uses $ICA_ACCESS_TOKEN env var).}
}
\value{
Tibble with file basename, file size, file full data path, file dir name.
}
\description{
List files on ICA GDS file system.
}
