% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{gds_files_list_filter_relevant}
\alias{gds_files_list_filter_relevant}
\title{List Relevant Files In GDS Directory}
\usage{
gds_files_list_filter_relevant(
  gdsdir,
  token,
  pattern = NULL,
  include_url = FALSE,
  page_size = 100,
  regexes = DR_FILE_REGEX
)
}
\arguments{
\item{gdsdir}{GDS directory.}

\item{token}{ICA access token.}

\item{pattern}{Pattern to further filter the returned file type tibble.}

\item{include_url}{Include presigned URLs to all files within the GDS directory (def: FALSE).}

\item{page_size}{Page size (def: 100).}

\item{regexes}{Tibble with regex and function name.}
}
\value{
A tibble with type, bname, size, file_id, path, and presigned URL.
}
\description{
Lists relevant files in a GDS directory.
}
