% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{gds_list_files_filter_relevant}
\alias{gds_list_files_filter_relevant}
\title{List Relevant Files In ICAv1 GDS Directory}
\usage{
gds_list_files_filter_relevant(
  gdsdir,
  pattern = NULL,
  regexes = DR_FILE_REGEX,
  token = Sys.getenv("ICA_ACCESS_TOKEN"),
  page_size = 100,
  include_url = FALSE,
  no_recurse = TRUE,
  page_token = NULL,
  recursive = NULL
)
}
\arguments{
\item{gdsdir}{Full path to GDS directory.}

\item{pattern}{Pattern to further filter the returned file type tibble.}

\item{regexes}{Tibble with \code{regex} and \code{fun}ction name.}

\item{token}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}

\item{page_size}{Page size (def: 10 via ICA API).}

\item{include_url}{Include presigned URLs to all files within the GDS directory (def: FALSE via ICA API).}

\item{no_recurse}{Do not recurse through the file list i.e. just give the first <page_size> items
without recursing further down the list using <page_token>.}

\item{page_token}{Page token (def: NULL). Used internally for recursion.}

\item{recursive}{Should files be returned recursively \emph{in and under} the specified
GDS directory, or \emph{only directly in} the specified GDS directory (def: TRUE via ICA API).}
}
\value{
A tibble with file type, basename, size, last modified timestamp, file_id, full path,
and presigned URL if requested.
}
\description{
Lists relevant files in a GDS directory.
}
\examples{
\dontrun{
regexes <- tibble::tibble(regex = "multiqc_data\\\\.json$", fun = "MultiqcJsonFile")
gdsdir <- "gds://production/analysis_data/SBJ01155/umccrise/202408300c218043/L2101566__L2101565"
gds_list_files_filter_relevant(gdsdir)
}
}
