% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal_meta.R
\name{meta_wgs_alignment_qc}
\alias{meta_wgs_alignment_qc}
\title{Metadata for wgs_alignment_qc workflow}
\usage{
meta_wgs_alignment_qc(pmeta, status = "Succeeded")
}
\arguments{
\item{pmeta}{Path to portal workflows metadata table, or tibble with already parsed data.}

\item{status}{Workflow status to keep (default: Succeeded).}
}
\value{
A tibble with metadata per workflow run.
}
\description{
Metadata for wgs_alignment_qc workflow
}
\examples{
\dontrun{
pmeta <- here::here("nogit/data_portal/2023-05-21_workflows.csv")
meta_wgs_alignment_qc(pmeta)
}
}
