% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiqc.R
\name{multiqc_parse_bargraph_plot}
\alias{multiqc_parse_bargraph_plot}
\title{MultiQC Extract Bar Graph Data}
\usage{
multiqc_parse_bargraph_plot(dat)
}
\arguments{
\item{dat}{Parsed JSON list element with specific plot data to extract.}
}
\value{
Tibble with name and data list-column.
}
\description{
Extracts \code{bar_graph} data for the given plot name from a MultiQC JSON.
}
