% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{PloidyEstimationMetricsFile}
\alias{PloidyEstimationMetricsFile}
\title{PloidyEstimationMetricsFile R6 Class}
\description{
Contains methods for reading contents of
the \code{ploidy_estimation_metrics.csv} file output from DRAGEN.
}
\examples{
x <- system.file("extdata/wgs/SEQC-II.ploidy_estimation_metrics.csv.gz", package = "dracarys")
pem <- PloidyEstimationMetricsFile$new(x)
pem$read() # or read(pem)

}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{PloidyEstimationMetricsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PloidyEstimationMetricsFile-read}{\code{PloidyEstimationMetricsFile$read()}}
\item \href{#method-PloidyEstimationMetricsFile-clone}{\code{PloidyEstimationMetricsFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="write"><a href='../../dracarys/html/File.html#method-File-write'><code>dracarys::File$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PloidyEstimationMetricsFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-PloidyEstimationMetricsFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{ploidy_estimation_metrics.csv} file output from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PloidyEstimationMetricsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
- label: sample label (inferred from file name)
- var: variable of interest (e.g. X median coverage)
- value: value for specific variable (e.g. X median coverage
variable with a value of  50)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PloidyEstimationMetricsFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PloidyEstimationMetricsFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PloidyEstimationMetricsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
