% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplayFile.R
\name{ReplayFile}
\alias{ReplayFile}
\title{ReplayFile R6 Class}
\description{
Contains methods for reading contents of
the \code{replay.json} file output from DRAGEN, which contains the DRAGEN command
line, parameters and version for the specific run.
}
\examples{
x <- system.file("extdata/wgs/SEQC-II-replay.json.gz", package = "dracarys")
r <- ReplayFile$new(x)
r$read() # or read(r)
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{ReplayFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-read}{\code{ReplayFile$read()}}
\item \href{#method-clone}{\code{ReplayFile$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname">}\href{../../dracarys/html/File.html#method-bname}{\code{dracarys::File$bname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize">}\href{../../dracarys/html/File.html#method-initialize}{\code{dracarys::File$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print">}\href{../../dracarys/html/File.html#method-print}{\code{dracarys::File$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type">}\href{../../dracarys/html/File.html#method-type}{\code{dracarys::File$type()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{replay.json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplayFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with the following elements:
\itemize{
\item \code{command_line}: character of DRAGEN command line used.
\item \code{dragen_config}: tibble of parameters used for the DRAGEN run.
\item \code{system}: tibble with dragen_version, nodename, and kernel_release.
\item \code{label}: character of sample label (inferred from file name)
\item \code{hash_table_build}: tibble with details about the DRAGEN hash table build.
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReplayFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
