% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_s3.R
\name{s3_list_files_dir}
\alias{s3_list_files_dir}
\title{List Objects in AWS S3 Directory}
\usage{
s3_list_files_dir(s3dir, max_objects = Inf)
}
\arguments{
\item{s3dir}{S3 directory.}

\item{max_objects}{Maximum objects returned. By default, iterates by 1000
objects at a time until the bucket/prefix is exhausted.}
}
\value{
A tibble with object basename, size, last modified timestamp, and
full S3 path.
}
\description{
Lists all objects in an S3 directory.
}
\examples{
\dontrun{
p1 <- "s3://project-data-491085415398-ap-southeast-2/byob-icav2"
p2 <- "project-tothill-main/cuppa/analysis/oncoanalyser-wgts-dna/202601077e38c827"
s3dir <- file.path(p1, p2)
s3_list_files_dir(s3dir, max_objects = 15)
}
}
