% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_s3.R
\name{s3_list_files_dir}
\alias{s3_list_files_dir}
\title{List Objects in AWS S3 Directory}
\usage{
s3_list_files_dir(s3dir, max_objects = 1000)
}
\arguments{
\item{s3dir}{S3 directory.}

\item{max_objects}{Maximum objects returned.}
}
\value{
A tibble with object basename, size, last modified timestamp, and
full S3 path.
}
\description{
Returns some or all (up to 1,000) of the objects in an S3 directory.
}
\examples{
\dontrun{
p1 <- "s3://org.umccr.data.oncoanalyser/analysis_data/SBJ05373/sash"
p2 <- "20240707becde493/L2401018_L2401017/SBJ05373_MDX240220"
s3dir <- file.path(p1, p2, "cancer_report/cancer_report_tables")
s3_list_files_dir(s3dir, max_objects = 15)
}
}
