% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_s3.R
\name{s3_list_files_filter_relevant}
\alias{s3_list_files_filter_relevant}
\title{List Relevant Files In AWS S3 Directory}
\usage{
s3_list_files_filter_relevant(
  s3dir,
  pattern = NULL,
  regexes = DR_FILE_REGEX,
  max_objects = 100,
  presign = FALSE,
  expiry_sec = 3600
)
}
\arguments{
\item{s3dir}{S3 directory.}

\item{pattern}{Pattern to further filter the returned file type tibble.}

\item{regexes}{Tibble with \code{regex} and \code{fun}ction name.}

\item{max_objects}{Maximum objects returned (def: 1000). For a recursive listing
use \code{Inf}.}

\item{presign}{Include presigned URLs (def: FALSE).}

\item{expiry_sec}{Number of seconds the presigned URL will be valid for (if generated).}
}
\value{
A tibble with file type, basename, size, last modified timestamp,
full path, and presigned URL if requested.
}
\description{
Lists relevant files in an AWS S3 directory.
}
\examples{
\dontrun{
p1 <- "s3://org.umccr.data.oncoanalyser/analysis_data/SBJ05373/sash"
p2 <- "20240707becde493/L2401018_L2401017/SBJ05373_MDX240220"
s3dir <- file.path(p1, p2)
regexes <- tibble::tibble(regex = "multiqc_data\\\\.json$", fun = "MultiqcJsonFile")
s3_list_files_filter_relevant(s3dir = s3dir, regexes = regexes, max_objects = 300)
}
}
