% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{s3_list_files_filter_relevant}
\alias{s3_list_files_filter_relevant}
\title{List Relevant Files In AWS S3 Directory}
\usage{
s3_list_files_filter_relevant(
  s3dir,
  pattern = NULL,
  max_objects = 100,
  presign = FALSE,
  expiry_sec = 3600,
  regexes = DR_FILE_REGEX
)
}
\arguments{
\item{s3dir}{S3 directory.}

\item{pattern}{Pattern to further filter the returned file type tibble.}

\item{max_objects}{The total number of objects to return.}

\item{presign}{Include presigned URLs (def: FALSE).}

\item{expiry_sec}{Number of seconds the presigned URL will be valid for (if generated).}

\item{regexes}{Tibble with \code{regex} and \code{fun}ction name.}
}
\value{
A tibble with file type, basename, size, last modified timestamp,
full path, and presigned URL if requested.
}
\description{
Lists relevant files in an AWS S3 directory.
}
\examples{
\dontrun{
s3dir <- "s3://umccr-primary-data-prod/cancer_report_tables"
s3_list_files_filter_relevant(s3dir = s3dir, presign = FALSE)
}
}
