% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_files}
\alias{tidy_files}
\title{Tidy Files}
\usage{
tidy_files(x)
}
\arguments{
\item{x}{Tibble with \code{fun}ction to parse the file and \code{localpath} to the file.}
}
\value{
Tibble with parsed data in a \code{data} list-column.
}
\description{
Tidy Files
}
\examples{
\dontrun{
p1 <- "~/icav1/g/production/analysis_data/SBJ01155/umccrise/202408300c218043"
p2 <- "L2101566__L2101565/SBJ01155__PRJ211091-qc_summary.tsv.gz"
p <- file.path(p1, p2)
x <- tibble::tibble(fun = "readr::read_tsv", localpath = p)
)
}

}
