% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_files}
\alias{tidy_files}
\title{Tidy Files}
\usage{
tidy_files(x, envir = parent.frame())
}
\arguments{
\item{x}{Tibble with \code{localpath} to file and the function \code{type} to parse it.
The function must return a tibble with a \code{name} column and the tidied \code{data}
as a list-column (see example).}

\item{envir}{the environment in which to evaluate the function e.g. use \code{self}
when using inside R6 classes.}
}
\value{
Tibble with parsed data in a \code{data} list-column.
}
\description{
Tidies files into a tibble with parsed data.
}
\examples{
\dontrun{
p1 <- "~/icav1/g/production/analysis_data/SBJ01155/umccrise/202408300c218043"
p2 <- "L2101566__L2101565/SBJ01155__PRJ211091/cancer_report_tables"
p <- file.path(p1, p2, "SBJ01155__PRJ211091-qc_summary.tsv.gz")
p_dl <- file.path(
  p1, "L2101566__L2101565/SBJ01155__PRJ211091/small_variants",
  "SBJ01155__PRJ211091-somatic-PASS.vcf.gz"
)
fun <- function(x) {
  d <- readr::read_tsv(x)
  tibble::tibble(name = "table1", data = list(d[]))
}
x <- tibble::tibble(
  type = c("fun", "DOWNLOAD_ONLY_foobar"), localpath = c(p, p_dl)
)
tidy_files(x)
}

}
