% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dragen.R
\name{TimeMetricsFile}
\alias{TimeMetricsFile}
\title{TimeMetricsFile R6 Class}
\description{
Contains methods for reading contents of
the \code{time_metrics.csv} file output from DRAGEN, which contains
a breakdown of the run duration for each DRAGEN process.
}
\examples{
x <- system.file("extdata/wgs/SEQC-II.time_metrics.csv.gz", package = "dracarys")
tm <- TimeMetricsFile$new(x)
tm$read() # or read(tm)
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TimeMetricsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeMetricsFile-read}{\code{TimeMetricsFile$read()}}
\item \href{#method-TimeMetricsFile-clone}{\code{TimeMetricsFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeMetricsFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TimeMetricsFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{time_metrics.csv} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeMetricsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item Step: DRAGEN step
\item Time: time in HH:MM
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeMetricsFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeMetricsFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeMetricsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
