% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso_acfc.R
\name{tso_acfc_read}
\alias{tso_acfc_read}
\title{Read TSO AlignCollapseFusionCaller_metrics File}
\usage{
tso_acfc_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\description{
Reads the \code{AlignCollapseFusionCaller_metrics.json.gz} file output from the
TSO500 (post-processing) workflow.
}
\examples{
x <- system.file("extdata/tso/sample705.AlignCollapseFusionCaller_metrics.json.gz",
  package = "dracarys"
)
d <- tso_acfc_read(x)
p <- tso_acfc_plot(d)
}
