% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{tso_combinedvaro_smallv_read}
\alias{tso_combinedvaro_smallv_read}
\title{Read TSO CombinedVariantOutput Small Variants}
\usage{
tso_combinedvaro_smallv_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\description{
Reads the \code{CombinedVariantOutput.tsv} file output from TSO and extracts
the Small Variants section (due to inconsistencies with other sections).
}
