% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{tso_fusions_read}
\alias{tso_fusions_read}
\title{Read TSO Fusions File}
\usage{
tso_fusions_read(x)
}
\arguments{
\item{x}{Path to file.}
}
\description{
Reads the \code{Fusions.csv} file output from the TSO500 workflow.
}
\examples{
x <- system.file("extdata/tso/sample705_Fusions.csv", package = "dracarys")
tso_fusions_read(x)
}
