% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{tso_rmd}
\alias{tso_rmd}
\title{Generate TSO Report}
\usage{
tso_rmd(prefix = NULL, out_file = NULL, quiet = FALSE)
}
\arguments{
\item{prefix}{Prefix for input files i.e. '/path/to/file1'.ext, where 'file1'
is the common prefix of all files.}

\item{out_file}{Name of output HTML file (needs '.html' suffix) (def: \verb{\{tumor_name\}_cancer_report.html}).}

\item{quiet}{Suppress warning and other messages.}
}
\value{
Path to rendered HTML report.
}
\description{
Generate TSO Report
}
