% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{tso_rmd}
\alias{tso_rmd}
\title{Generate TSO Report}
\usage{
tso_rmd(tidy_rds, out_dir, prefix, quiet = FALSE)
}
\arguments{
\item{tidy_rds}{Path to RDS object output from \code{tso_tidy}.}

\item{out_dir}{Output directory.}

\item{prefix}{Prefix of output file(s).}

\item{quiet}{Suppress warning and other messages.}
}
\value{
Path to rendered HTML report.
}
\description{
Generates a HTML report with results from the UMCCR TSO500 ctDNA workflow.
}
\examples{
\dontrun{
tidy_rds <- "~/tmp/tso/SBJ02858.rds"
tso_tidy(...) |> saveRDS(tidy_rds)
quiet <- TRUE
out_dir <- file.path("~/tmp/tso/out2")
prefix <- "SBJ02858"
tso_rmd(tidy_rds = tidy_rds, out_dir = out_dir, prefix = prefix, quiet = quiet)
}
}
