% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{tso_tidy}
\alias{tso_tidy}
\title{Tidy TSO ctDNA Results}
\usage{
tso_tidy(
  indir,
  outprefix,
  gds_local_dir = NULL,
  out_format = "tsv",
  dryrun = FALSE,
  token = Sys.getenv("ICA_ACCESS_TOKEN")
)
}
\arguments{
\item{indir}{Directory path to TSO500 ctDNA workflow results (must end
with \code{/}).}

\item{outprefix}{Prefix path of output file(s).}

\item{gds_local_dir}{If \code{indir} is a GDS directory, 'recognisable' files
will be first downloaded to this directory.}

\item{out_format}{Format of output (tsv, parquet, both) (def: tsv).}

\item{dryrun}{Just list the files that will be downloaded?}

\item{token}{ICA access token (by default uses $ICA_ACCESS_TOKEN env var).}
}
\value{
List of tibbles.
}
\description{
Tidies TSO ctDNA results into a list of tibbles.
}
\examples{
\dontrun{
indir <- paste0(
  "gds://production/analysis_data/SBJ02858/tso_ctdna_tumor_only/",
  "20221104b7ad0b38/L2201560/Results/PRJ222206_L2201560/"
)
indir <- here::here(glue("nogit/tso/2022-12-13/SBJ02858/dracarys_gds_sync"))
gds_local_dir <- NULL
outprefix <- file.path(indir, "out/SBJ02858")
dryrun <- F
tso_tidy(indir = indir, outprefix = outprefix)
}
}
