% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso_acfc.R
\name{TsoAlignCollapseFusionCallerMetricsFile}
\alias{TsoAlignCollapseFusionCallerMetricsFile}
\title{TsoAlignCollapseFusionCallerMetricsFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{AlignCollapseFusionCaller_metrics.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705.AlignCollapseFusionCaller_metrics.json.gz",
  package = "dracarys"
)
m <- TsoAlignCollapseFusionCallerMetricsFile$new(x)
d_parsed <- m$read() # or read(m)
m$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = c("tsv", "rds"))
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoAlignCollapseFusionCallerMetricsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoAlignCollapseFusionCallerMetricsFile-read}{\code{TsoAlignCollapseFusionCallerMetricsFile$read()}}
\item \href{#method-TsoAlignCollapseFusionCallerMetricsFile-write}{\code{TsoAlignCollapseFusionCallerMetricsFile$write()}}
\item \href{#method-TsoAlignCollapseFusionCallerMetricsFile-plot}{\code{TsoAlignCollapseFusionCallerMetricsFile$plot()}}
\item \href{#method-TsoAlignCollapseFusionCallerMetricsFile-clone}{\code{TsoAlignCollapseFusionCallerMetricsFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoAlignCollapseFusionCallerMetricsFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoAlignCollapseFusionCallerMetricsFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{AlignCollapseFusionCaller_metrics.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoAlignCollapseFusionCallerMetricsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item section: name of original JSON element
\item name: name of metric
\item value: value of metric
\item percent: percentage
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoAlignCollapseFusionCallerMetricsFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TsoAlignCollapseFusionCallerMetricsFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{AlignCollapseFusionCaller_metrics.json.gz} file
output from TSO.

Histo is the majority from UmiStatistics section, write out separately.
Histo of num supporting fragments: Num of families with 0/1/2/3... raw reads.
Histo of unique UMIs per fragment pos: Num of pos with 0/1/2/3... UMI seqs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoAlignCollapseFusionCallerMetricsFile$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s) (one of 'tsv' (def.),
'parquet', 'both').}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoAlignCollapseFusionCallerMetricsFile-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TsoAlignCollapseFusionCallerMetricsFile-plot}{}}}
\subsection{Method \code{plot()}}{
Generates the UmiStatistics Histogram plots from the
\code{AlignCollapseFusionCaller_metrics.json.gz} file output from TSO.
\itemize{
\item Histo is the majority from UmiStatistics section, deal with it separately.
\item Histo of num supporting fragments: Num of families with 0/1/2/3... raw reads.
\item Histo of unique UMIs per fragment pos: Num of pos with 0/1/2/3... UMI seqs.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoAlignCollapseFusionCallerMetricsFile$plot(d, max_num = 15)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{max_num}}{Maximum number to display in both plots.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Both histogram plot objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoAlignCollapseFusionCallerMetricsFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoAlignCollapseFusionCallerMetricsFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoAlignCollapseFusionCallerMetricsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
