% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoAlignCollapseFusionCallerMetricsFile}
\alias{TsoAlignCollapseFusionCallerMetricsFile}
\title{TsoAlignCollapseFusionCallerMetricsFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{AlignCollapseFusionCaller_metrics.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705.AlignCollapseFusionCaller_metrics.json.gz", package = "dracarys")
m <- TsoAlignCollapseFusionCallerMetricsFile$new(x)
m$read() # or read(m)
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoAlignCollapseFusionCallerMetricsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoAlignCollapseFusionCallerMetricsFile-read}{\code{TsoAlignCollapseFusionCallerMetricsFile$read()}}
\item \href{#method-TsoAlignCollapseFusionCallerMetricsFile-clone}{\code{TsoAlignCollapseFusionCallerMetricsFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoAlignCollapseFusionCallerMetricsFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoAlignCollapseFusionCallerMetricsFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{AlignCollapseFusionCaller_metrics.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoAlignCollapseFusionCallerMetricsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item section: name of original JSON element
\item name: name of metric
\item value: value of metric
\item percent: percentage
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoAlignCollapseFusionCallerMetricsFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoAlignCollapseFusionCallerMetricsFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoAlignCollapseFusionCallerMetricsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
