% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoCombinedVariantOutputFile}
\alias{TsoCombinedVariantOutputFile}
\title{TsoCombinedVariantOutputFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{CombinedVariantOutput.tsv} file output from TSO.
}
\examples{
\dontrun{
x <- "/path/to/CombinedVariantOutput.tsv"
d <- TsoCombinedVariantOutputFile$new(x)
d$read()
}
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoCombinedVariantOutputFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoCombinedVariantOutputFile-read}{\code{TsoCombinedVariantOutputFile$read()}}
\item \href{#method-TsoCombinedVariantOutputFile-write}{\code{TsoCombinedVariantOutputFile$write()}}
\item \href{#method-TsoCombinedVariantOutputFile-clone}{\code{TsoCombinedVariantOutputFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoCombinedVariantOutputFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoCombinedVariantOutputFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{CombinedVariantOutput.tsv} file output from TSO and extracts
only the Small Variants section (due to inconsistencies with other sections).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoCombinedVariantOutputFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with variants.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoCombinedVariantOutputFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TsoCombinedVariantOutputFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{CombinedVariantOutput.tsv} (only Small Variants)
file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoCombinedVariantOutputFile$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoCombinedVariantOutputFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoCombinedVariantOutputFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoCombinedVariantOutputFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
