% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoFragmentLengthHistFile}
\alias{TsoFragmentLengthHistFile}
\title{TsoFragmentLengthHistFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{fragment_length_hist.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705.fragment_length_hist.json.gz", package = "dracarys")
fl <- TsoFragmentLengthHistFile$new(x)
d_parsed <- fl$read() # or read(fl)
fl$plot(d_parsed, 5)
fl$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "both")
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoFragmentLengthHistFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoFragmentLengthHistFile-read}{\code{TsoFragmentLengthHistFile$read()}}
\item \href{#method-TsoFragmentLengthHistFile-write}{\code{TsoFragmentLengthHistFile$write()}}
\item \href{#method-TsoFragmentLengthHistFile-plot}{\code{TsoFragmentLengthHistFile$plot()}}
\item \href{#method-TsoFragmentLengthHistFile-clone}{\code{TsoFragmentLengthHistFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFragmentLengthHistFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFragmentLengthHistFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{fragment_length_hist.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFragmentLengthHistFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item FragmentLength
\item Count
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFragmentLengthHistFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFragmentLengthHistFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{fragment_length_hist.json.gz} file output
from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFragmentLengthHistFile$write(d, out_dir, prefix, out_format = "tsv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s) (one of 'tsv' (def.),
'parquet', 'both').}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFragmentLengthHistFile-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFragmentLengthHistFile-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the fragment length distributions as given in the
\code{fragment_length_hist.json.gz} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFragmentLengthHistFile$plot(d, min_count = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{min_count}}{Minimum read count to be plotted (Default: 10).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 plot containing fragment lengths on X axis and read counts
on Y axis for each sample.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFragmentLengthHistFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFragmentLengthHistFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFragmentLengthHistFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
