% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoFusionsCsvFile}
\alias{TsoFusionsCsvFile}
\title{TsoFusionsCsvFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{Fusions.csv} file output from TSO.
Returns a tibble where the columns are (based on the input file header metadata):
\itemize{
\item Sample: input sample ID.
\item Name: Fusion name as reported by manta.
\item Chr1/Chr2: The chromosome of the 1st/2nd breakend.
\item Pos1/Pos2: The position of the 1st/2nd breakend.
\item Direction: The direction of how the breakends are joined together.
\item Alt_Depth: The number of read-pairs supporting the fusion call.
\item BP1_Depth/BP2_Depth: Number of read-pairs aligned to the 1st/2nd breakend.
\item Total_Depth: Max number of read-pairs aligned to a fusion breakend.
\item VAF: Variant allele frequency.
\item Gene1/Gene2: Genes that overlap the 1st/2nd breakend.
\item Contig: The fusion contig.
\item Filter: Indicates whether the fusion has passed all of the fusion filters.
\item Is_Cosmic_GenePair: Indicates whether the gene pair has been reported by Cosmic(True/False).
}
}
\examples{
x <- system.file("extdata/tso/sample705_Fusions.csv", package = "dracarys")
fus <- TsoFusionsCsvFile$new(x)
d_parsed <- fus$read() # or read(fus)
fus$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "tsv")
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoFusionsCsvFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoFusionsCsvFile-read}{\code{TsoFusionsCsvFile$read()}}
\item \href{#method-TsoFusionsCsvFile-write}{\code{TsoFusionsCsvFile$write()}}
\item \href{#method-TsoFusionsCsvFile-clone}{\code{TsoFusionsCsvFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFusionsCsvFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFusionsCsvFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{Fusions.csv} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFusionsCsvFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with several columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFusionsCsvFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFusionsCsvFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{Fusions.csv} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFusionsCsvFile$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFusionsCsvFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFusionsCsvFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFusionsCsvFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
