% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoFusionsCsvFile}
\alias{TsoFusionsCsvFile}
\title{TsoFusionsCsvFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{Fusions.csv} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705_Fusions.csv", package = "dracarys")
fus <- TsoFusionsCsvFile$new(x)
fus$read() # or read(fus)
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoFusionsCsvFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoFusionsCsvFile-read}{\code{TsoFusionsCsvFile$read()}}
\item \href{#method-TsoFusionsCsvFile-clone}{\code{TsoFusionsCsvFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFusionsCsvFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFusionsCsvFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{Fusions.csv} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFusionsCsvFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item label:
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoFusionsCsvFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoFusionsCsvFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoFusionsCsvFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
