% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoMsiFile}
\alias{TsoMsiFile}
\title{TsoMsiFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{msi.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705.msi.json.gz", package = "dracarys")
msi <- TsoMsiFile$new(x)
msi$read() # or read(msi)
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoMsiFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoMsiFile-read}{\code{TsoMsiFile$read()}}
\item \href{#method-TsoMsiFile-clone}{\code{TsoMsiFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoMsiFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoMsiFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{msi.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoMsiFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item label:
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoMsiFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoMsiFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoMsiFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
