% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoSampleAnalysisResultsFile}
\alias{TsoSampleAnalysisResultsFile}
\title{TsoSampleAnalysisResultsFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{SampleAnalysisResults.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705_SampleAnalysisResults.json.gz", package = "dracarys")
res <- TsoSampleAnalysisResultsFile$new(x)
res$read() # or read(res)
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoMsiFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-read}{\code{TsoSampleAnalysisResultsFile$read()}}
\item \href{#method-clone}{\code{TsoSampleAnalysisResultsFile$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname">}\href{../../dracarys/html/File.html#method-bname}{\code{dracarys::File$bname()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize">}\href{../../dracarys/html/File.html#method-initialize}{\code{dracarys::File$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print">}\href{../../dracarys/html/File.html#method-print}{\code{dracarys::File$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type">}\href{../../dracarys/html/File.html#method-type}{\code{dracarys::File$type()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{msi.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoSampleAnalysisResultsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
- TOTAL_PF_READS
- MEAN_FAMILY_SIZE
- MEDIAN_TARGET_COVERAGE,
- PCT_CHIMERIC_READS
- PCT_EXON_500X
- PCT_EXON_1500X
- PCT_READ_ENRICHMENT,
- PCT_USABLE_UMI_READS
- MEAN_TARGET_COVERAGE
- PCT_ALIGNED_READS,
- PCT_CONTAMINATION_EST
- PCT_TARGET_0.4X_MEAN
- PCT_TARGET_500X,
- PCT_TARGET_1000X
- PCT_TARGET_1500X
- PCT_DUPLEXFAMILIES,
- MEDIAN_INSERT_SIZE
- MAX_SOMATIC_AF
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoSampleAnalysisResultsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
