% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoSampleAnalysisResultsFile}
\alias{TsoSampleAnalysisResultsFile}
\title{TsoSampleAnalysisResultsFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{SampleAnalysisResults.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705_SampleAnalysisResults.json.gz", package = "dracarys")
res <- TsoSampleAnalysisResultsFile$new(x)
d_parsed <- res$read() # or read(res)
res$write(d_parsed, tempfile(), "both")
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoSampleAnalysisResultsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoSampleAnalysisResultsFile-read}{\code{TsoSampleAnalysisResultsFile$read()}}
\item \href{#method-TsoSampleAnalysisResultsFile-write}{\code{TsoSampleAnalysisResultsFile$write()}}
\item \href{#method-TsoSampleAnalysisResultsFile-clone}{\code{TsoSampleAnalysisResultsFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoSampleAnalysisResultsFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoSampleAnalysisResultsFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{SampleAnalysisResults.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoSampleAnalysisResultsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of tibbles
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoSampleAnalysisResultsFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TsoSampleAnalysisResultsFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{SampleAnalysisResults.json.gz} file output
from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoSampleAnalysisResultsFile$write(d, out_dir, prefix, out_format = "tsv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s) (one of 'tsv' (def.),
'parquet', 'both').}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoSampleAnalysisResultsFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoSampleAnalysisResultsFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoSampleAnalysisResultsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
