% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoTargetRegionCoverageFile}
\alias{TsoTargetRegionCoverageFile}
\title{TsoTargetRegionCoverageFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{TargetRegionCoverage.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705.TargetRegionCoverage.json.gz", package = "dracarys")
trc <- TsoTargetRegionCoverageFile$new(x)
d_parsed <- trc$read() # or read(trc)
trc$plot(d_parsed, 90) # or plot(trc, d_parsed, 90)
trc$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "tsv")
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoTargetRegionCoverageFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoTargetRegionCoverageFile-read}{\code{TsoTargetRegionCoverageFile$read()}}
\item \href{#method-TsoTargetRegionCoverageFile-write}{\code{TsoTargetRegionCoverageFile$write()}}
\item \href{#method-TsoTargetRegionCoverageFile-plot}{\code{TsoTargetRegionCoverageFile$plot()}}
\item \href{#method-TsoTargetRegionCoverageFile-clone}{\code{TsoTargetRegionCoverageFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoTargetRegionCoverageFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoTargetRegionCoverageFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{TargetRegionCoverage.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoTargetRegionCoverageFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item ConsensusReadDepth
\item BasePair
\item Percentage
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoTargetRegionCoverageFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TsoTargetRegionCoverageFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{TargetRegionCoverage.json.gz} file output
from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoTargetRegionCoverageFile$write(
  d,
  out_dir = NULL,
  prefix,
  out_format = "tsv",
  drid = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s).}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoTargetRegionCoverageFile-plot"></a>}}
\if{latex}{\out{\hypertarget{method-TsoTargetRegionCoverageFile-plot}{}}}
\subsection{Method \code{plot()}}{
Plots the \code{TargetRegionCoverage.json.gz} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoTargetRegionCoverageFile$plot(d, min_pct = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{min_pct}}{Minimum percentage to be plotted (Default: 2).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot2 plot containing read depth on X axis and percentage
covered on Y axis.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoTargetRegionCoverageFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoTargetRegionCoverageFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoTargetRegionCoverageFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
