% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tso.R
\name{TsoTmbFile}
\alias{TsoTmbFile}
\title{TsoTmbFile R6 Class}
\description{
Contains methods for reading and displaying contents of the
\code{tmb.json.gz} file output from TSO.
}
\examples{
x <- system.file("extdata/tso/sample705.tmb.json.gz", package = "dracarys")
tmb <- TsoTmbFile$new(x)
d_parsed <- tmb$read() # or read(tmb)
tmb$write(d_parsed, out_dir = tempdir(), prefix = "sample705", out_format = "both")
}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{TsoTmbFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TsoTmbFile-read}{\code{TsoTmbFile$read()}}
\item \href{#method-TsoTmbFile-write}{\code{TsoTmbFile$write()}}
\item \href{#method-TsoTmbFile-clone}{\code{TsoTmbFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoTmbFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-TsoTmbFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{tmb.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoTmbFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item TmbPerMb
\item AdjustedTmbPerMb
\item NonsynonymousTmbPerMb
\item AdjustedNonsynonymousTmbPerMb
\item SomaticCodingVariantsCount
\item NonsynonymousSomaticCodingVariantsCount
\item TotalRegionSizeMb
\item CodingRegionSizeMb
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoTmbFile-write"></a>}}
\if{latex}{\out{\hypertarget{method-TsoTmbFile-write}{}}}
\subsection{Method \code{write()}}{
Writes a tidy version of the \code{tmb.json.gz} file output from TSO.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoTmbFile$write(d, out_dir = NULL, prefix, out_format = "tsv", drid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{d}}{Parsed object from \code{self$read()}.}

\item{\code{out_dir}}{Output directory.}

\item{\code{prefix}}{Prefix of output file(s).}

\item{\code{out_format}}{Format of output file(s) (one of 'tsv' (def.),
'parquet', 'both').}

\item{\code{drid}}{dracarys ID to use for the dataset (e.g. \code{wfrid.123}, \code{prid.456}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TsoTmbFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TsoTmbFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TsoTmbFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
