% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{umccr_tidy}
\alias{umccr_tidy}
\title{Tidy UMCCR Results}
\usage{
umccr_tidy(
  in_dir,
  out_dir,
  prefix,
  gds_local_dir = NULL,
  out_format = "tsv",
  dryrun = FALSE,
  token = Sys.getenv("ICA_ACCESS_TOKEN")
)
}
\arguments{
\item{in_dir}{Directory path to UMCCR workflow results (can be GDS or local).}

\item{out_dir}{Output directory.}

\item{prefix}{Prefix of output file(s).}

\item{gds_local_dir}{If \code{indir} is a GDS directory, 'recognisable' files
will be first downloaded to this directory.}

\item{out_format}{Format of output (tsv, parquet, both) (def: tsv).}

\item{dryrun}{Just list the files that will be downloaded (def: FALSE).}

\item{token}{ICA access token (by default uses $ICA_ACCESS_TOKEN env var).}
}
\value{
Tibble with path to input file and the resultant tidy object.
}
\description{
Tidies UMCCR workflow results into a list of tibbles and writes individual
tibbles to TSV and/or Parquet format.
}
\examples{
\dontrun{
in_dir <- paste0(
  "gds://production/analysis_data/SBJ02858/tso_ctdna_tumor_only/",
  "20221104b7ad0b38/L2201560/Results/PRJ222206_L2201560/"
)
in_dir <- here::here(glue("nogit/tso/2022-12-13/SBJ02858/dracarys_gds_sync"))
out_dir <- file.path(in_dir, "../out")
gds_local_dir <- NULL
prefix <- "SBJ02858"
dryrun <- F
umccr_tidy(in_dir = in_dir, out_dir = out_dir, prefix = prefix)
}
}
