% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{umccr_tidy}
\alias{umccr_tidy}
\title{Tidy UMCCR Results}
\usage{
umccr_tidy(
  in_dir = NULL,
  out_dir = NULL,
  prefix = NULL,
  gds_local_dir = NULL,
  out_format = "tsv",
  dryrun = FALSE,
  token = Sys.getenv("ICA_ACCESS_TOKEN"),
  pattern = NULL
)
}
\arguments{
\item{in_dir}{Directory path to UMCCR workflow results (can be GDS or local).}

\item{out_dir}{Output directory.}

\item{prefix}{Prefix of output file(s).}

\item{gds_local_dir}{If \code{indir} is a GDS directory, 'recognisable' files
will be first downloaded to this directory.}

\item{out_format}{Format of output (tsv, parquet, both) (def: tsv).}

\item{dryrun}{Just list the files that will be downloaded (def: FALSE).}

\item{token}{ICA access token (by default uses $ICA_ACCESS_TOKEN env var).}

\item{pattern}{Pattern to further filter the returned file type tibble (see
\code{name} column in the \code{DR_FILE_REGEX} tibble).}
}
\value{
Tibble with path to input file and the resultant tidy object.
}
\description{
Tidies UMCCR workflow results into a list of tibbles and writes individual
tibbles to TSV, Parquet, SparkDF, or RDS format.
}
\examples{
\dontrun{
in_dir <- paste0(
  "gds://production/analysis_data/SBJ01639/tso_ctdna_tumor_only/",
  "202204045ad5743c/L2200214/Results/PRJ220425_L2200214"
)
o1 <- sub("^gds://", "", in_dir)
out_dir <- file.path(fs::path_home(), "icav1/g", o1)
# in_dir <- file.path(out_dir, "dracarys_gds_sync")
prefix <- "SBJ01639"
out_format <- "rds"
umccr_tidy(in_dir = in_dir, out_dir = out_dir, prefix = prefix, out_format = out_format, dryrun = T)

in_dir <- here::here(glue("nogit/tso/2022-12-13/SBJ02858/dracarys_gds_sync"))
out_dir <- file.path(in_dir, "../out")
gds_local_dir <- NULL
prefix <- "SBJ02858"
dryrun <- F
umccr_tidy(in_dir = in_dir, out_dir = out_dir, prefix = prefix, dryrun = F)
}
}
