% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCMetricsFile.R
\name{VCMetricsFile}
\alias{VCMetricsFile}
\title{VCMetricsFile R6 Class}
\description{
Contains methods for reading and displaying contents of
the \code{vc_metrics.csv} file output from DRAGEN, which contains variant calling metrics.
reported for each sample in multi sample VCF and gVCF files.
Based on the run case, metrics are reported either as standard VARIANT
CALLER or JOINT CALLER. Metrics are reported both for the raw
(PREFILTER) and hard filtered (POSTFILTER) VCFs.
PON (Panel of Normals) and COSMIC filtered variants are counted as
though they are PASS variants in the POSTFILTER VCFs metrics,
which may result in higher than expected variant counts in the
POSTFILTER VCF metrics.
}
\examples{
x <- system.file("extdata/wgs/SEQC-II.vc_metrics.csv.gz", package = "dracarys")
vm <- VCMetricsFile$new(x)
vm$read() # or read(vm)

}
\section{Super class}{
\code{\link[dracarys:File]{dracarys::File}} -> \code{VCMetricsFile}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VCMetricsFile-read}{\code{VCMetricsFile$read()}}
\item \href{#method-VCMetricsFile-clone}{\code{VCMetricsFile$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="bname"><a href='../../dracarys/html/File.html#method-File-bname'><code>dracarys::File$bname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="initialize"><a href='../../dracarys/html/File.html#method-File-initialize'><code>dracarys::File$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="print"><a href='../../dracarys/html/File.html#method-File-print'><code>dracarys::File$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="File" data-id="type"><a href='../../dracarys/html/File.html#method-File-type'><code>dracarys::File$type()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCMetricsFile-read"></a>}}
\if{latex}{\out{\hypertarget{method-VCMetricsFile-read}{}}}
\subsection{Method \code{read()}}{
Reads the \code{vc_metrics.csv} file output from DRAGEN.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCMetricsFile$read()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble with the following columns:
\itemize{
\item category
\item sample
\item var: variable name
\item count: count value
\item pct: percent value
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VCMetricsFile-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VCMetricsFile-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VCMetricsFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
