% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sash.R
\name{Wf_sash}
\alias{Wf_sash}
\title{Wf_sash R6 Class}
\description{
Reads and writes tidy versions of files from the \code{sash} workflow
}
\examples{
\dontrun{

#---- Local ----#
p1 <- "~/s3/project-data-889522050439-ap-southeast-2/byob-icav2"
p2 <- "project-wgs-accreditation/analysis/sash/20250919121765d8"
p <- normalizePath(file.path(p1, p2))
libid_tumor <- "L2300943"
libid_normal <- "L2300950"
s1 <- Wf_sash$new(path = p, libid_tumor = libid_tumor, libid_normal = libid_normal)
s1$list_files(max_files = 20)
s1$list_files_filter_relevant(max_files = 300)
d <- s1$download_files(max_files = 1000, dryrun = F)
d_tidy <- s1$tidy_files(d)
d_write <- s1$write(
  d_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = glue("{libid_normal}_{libid_tumor}"),
  format = "tsv"
)

#---- S3 ----#
p1 <- "s3://org.umccr.data.oncoanalyser/analysis_data/SBJ02862/sash"
p2 <- "20240830ece6b0b7/L2201449_L2201450"
p <- file.path(p1, p2)
libid_tumor <- "PRJ222112"
libid_normal <- "PRJ222114"
s1 <- Wf_sash$new(
  path = p,
  libid_tumor = libid_tumor, libid_normal = libid_normal
)
s1$list_files(max_files = 20)
s1$list_files_filter_relevant()
outdir <- sub("s3:/", "~/s3", p)
d <- s1$download_files(outdir = outdir, max_files = 1000, dryrun = F)
d_tidy <- s1$tidy_files(d)
d_write <- s1$write(
  d_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = glue("__{libid_tumor}"),
  format = "tsv"
)
}

}
\section{Super class}{
\code{\link[dracarys:Wf]{dracarys::Wf}} -> \code{Wf_sash}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{libid_tumor}}{The LibraryID of the tumor sample.}

\item{\code{libid_normal}}{The LibraryID of the normal sample.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf_sash-new}{\code{Wf_sash$new()}}
\item \href{#method-Wf_sash-print}{\code{Wf_sash$print()}}
\item \href{#method-Wf_sash-read_smlvSomCounts}{\code{Wf_sash$read_smlvSomCounts()}}
\item \href{#method-Wf_sash-read_pcgrJson}{\code{Wf_sash$read_pcgrJson()}}
\item \href{#method-Wf_sash-read_hrdDragen}{\code{Wf_sash$read_hrdDragen()}}
\item \href{#method-Wf_sash-read_hrdChord}{\code{Wf_sash$read_hrdChord()}}
\item \href{#method-Wf_sash-read_hrdHrdetect}{\code{Wf_sash$read_hrdHrdetect()}}
\item \href{#method-Wf_sash-read_sigsTsv}{\code{Wf_sash$read_sigsTsv()}}
\item \href{#method-Wf_sash-read_qcSum}{\code{Wf_sash$read_qcSum()}}
\item \href{#method-Wf_sash-clone}{\code{Wf_sash$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="DOWNLOAD_ONLY"><a href='../../dracarys/html/Wf.html#method-Wf-DOWNLOAD_ONLY'><code>dracarys::Wf$DOWNLOAD_ONLY()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="download_files"><a href='../../dracarys/html/Wf.html#method-Wf-download_files'><code>dracarys::Wf$download_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files"><a href='../../dracarys/html/Wf.html#method-Wf-list_files'><code>dracarys::Wf$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files_filter_relevant"><a href='../../dracarys/html/Wf.html#method-Wf-list_files_filter_relevant'><code>dracarys::Wf$list_files_filter_relevant()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="tidy_files"><a href='../../dracarys/html/Wf.html#method-Wf-tidy_files'><code>dracarys::Wf$tidy_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="write"><a href='../../dracarys/html/Wf.html#method-Wf-write'><code>dracarys::Wf$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-new"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-new}{}}}
\subsection{Method \code{new()}}{
Create a new Wf_sash object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$new(path = NULL, libid_tumor = NULL, libid_normal = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to directory with raw workflow results (from S3 or
local filesystem).}

\item{\code{libid_tumor}}{The LibraryID of the tumor sample.}

\item{\code{libid_normal}}{The LibraryID of the normal sample.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-print"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-print}{}}}
\subsection{Method \code{print()}}{
Print details about the Workflow.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_smlvSomCounts"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_smlvSomCounts}{}}}
\subsection{Method \code{read_smlvSomCounts()}}{
Read \code{somatic.variant_counts_process.json} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_smlvSomCounts(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_pcgrJson"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_pcgrJson}{}}}
\subsection{Method \code{read_pcgrJson()}}{
Read \code{pcgr.json.gz} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_pcgrJson(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_hrdDragen"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_hrdDragen}{}}}
\subsection{Method \code{read_hrdDragen()}}{
Read \code{dragen.tsv.gz} cancer report hrd file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_hrdDragen(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_hrdChord"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_hrdChord}{}}}
\subsection{Method \code{read_hrdChord()}}{
Read \code{chord.tsv.gz} cancer report hrd file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_hrdChord(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_hrdHrdetect"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_hrdHrdetect}{}}}
\subsection{Method \code{read_hrdHrdetect()}}{
Read \code{hrdetect.tsv.gz} cancer report hrd file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_hrdHrdetect(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_sigsTsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_sigsTsv}{}}}
\subsection{Method \code{read_sigsTsv()}}{
Read signature cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_sigsTsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_qcSum"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_qcSum}{}}}
\subsection{Method \code{read_qcSum()}}{
Read \code{qc_summary.tsv.gz} cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_qcSum(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
