% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sash.R
\name{Wf_sash}
\alias{Wf_sash}
\title{Wf_sash R6 Class}
\description{
Reads and writes tidy versions of files from the \code{sash} workflow
}
\examples{
\dontrun{

#---- Local ----#
p1 <- "~/s3/org.umccr.data.oncoanalyser/analysis_data/SBJ03324/sash"
p2 <- "202408309698c304/L2300777_L2300776"
p <- normalizePath(file.path(p1, p2))
SubjectID <- "SBJ03324"
SampleID_tumor <- "PRJ230432"
prefix <- glue("{SubjectID}__{SampleID_tumor}")
s1 <- Wf_sash$new(path = p, SubjectID = SubjectID, SampleID_tumor = SampleID_tumor)
s1$list_files(max_files = 20)
s1$list_files_filter_relevant(max_files = 300)
d <- s1$download_files(max_files = 1000, dryrun = F)
d_tidy <- s1$tidy_files(d)
d_write <- s1$write(
  d_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = glue("{SubjectID}_{SampleID_tumor}"),
  format = "tsv"
)

#---- S3 ----#
p1 <- "s3://org.umccr.data.oncoanalyser/analysis_data/SBJ05571/sash"
p2 <- "202408270b93455e/L2401308_L2401307"
p <- file.path(p1, p2)
SubjectID <- "SBJ05571"
SampleID_tumor <- "MDX240307"
prefix <- glue("{SubjectID}__{SampleID_tumor}")
s1 <- Wf_sash$new(path = p, SubjectID = SubjectID, SampleID_tumor = SampleID_tumor)
s1$list_files(max_files = 20)
s1$list_files_filter_relevant()
outdir <- sub("s3:/", "~/s3", p)
d <- s1$download_files(outdir = outdir, max_files = 1000, dryrun = F)
d_tidy <- s1$tidy_files(d)
d_write <- s1$write(
  d_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = glue("{SubjectID}__{SampleID_tumor}"),
  format = "tsv"
)
}

}
\section{Super class}{
\code{\link[dracarys:Wf]{dracarys::Wf}} -> \code{Wf_sash}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{SubjectID}}{The SubjectID of the sample (needed for path lookup).}

\item{\code{SampleID_tumor}}{The SampleID of the tumor sample (needed for path lookup).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf_sash-new}{\code{Wf_sash$new()}}
\item \href{#method-Wf_sash-print}{\code{Wf_sash$print()}}
\item \href{#method-Wf_sash-read_pcgr_json}{\code{Wf_sash$read_pcgr_json()}}
\item \href{#method-Wf_sash-read_hrd_dragen}{\code{Wf_sash$read_hrd_dragen()}}
\item \href{#method-Wf_sash-read_hrd_chord}{\code{Wf_sash$read_hrd_chord()}}
\item \href{#method-Wf_sash-read_hrd_hrdetect}{\code{Wf_sash$read_hrd_hrdetect()}}
\item \href{#method-Wf_sash-read_sigstsv}{\code{Wf_sash$read_sigstsv()}}
\item \href{#method-Wf_sash-read_qcsum}{\code{Wf_sash$read_qcsum()}}
\item \href{#method-Wf_sash-clone}{\code{Wf_sash$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="download_files"><a href='../../dracarys/html/Wf.html#method-Wf-download_files'><code>dracarys::Wf$download_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files"><a href='../../dracarys/html/Wf.html#method-Wf-list_files'><code>dracarys::Wf$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files_filter_relevant"><a href='../../dracarys/html/Wf.html#method-Wf-list_files_filter_relevant'><code>dracarys::Wf$list_files_filter_relevant()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="tidy_files"><a href='../../dracarys/html/Wf.html#method-Wf-tidy_files'><code>dracarys::Wf$tidy_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="write"><a href='../../dracarys/html/Wf.html#method-Wf-write'><code>dracarys::Wf$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-new"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-new}{}}}
\subsection{Method \code{new()}}{
Create a new Wf_sash object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$new(path = NULL, SubjectID = NULL, SampleID_tumor = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to directory with raw workflow results (from GDS, S3, or
local filesystem).}

\item{\code{SubjectID}}{The SubjectID of the sample (needed for path lookup).}

\item{\code{SampleID_tumor}}{The SampleID of the tumor sample (needed for path lookup).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-print"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-print}{}}}
\subsection{Method \code{print()}}{
Print details about the Workflow.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_pcgr_json"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_pcgr_json}{}}}
\subsection{Method \code{read_pcgr_json()}}{
Read \code{pcgr.json.gz} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_pcgr_json(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_hrd_dragen"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_hrd_dragen}{}}}
\subsection{Method \code{read_hrd_dragen()}}{
Read \code{dragen.tsv.gz} cancer report hrd file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_hrd_dragen(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_hrd_chord"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_hrd_chord}{}}}
\subsection{Method \code{read_hrd_chord()}}{
Read \code{chord.tsv.gz} cancer report hrd file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_hrd_chord(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_hrd_hrdetect"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_hrd_hrdetect}{}}}
\subsection{Method \code{read_hrd_hrdetect()}}{
Read \code{hrdetect.tsv.gz} cancer report hrd file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_hrd_hrdetect(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_sigstsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_sigstsv}{}}}
\subsection{Method \code{read_sigstsv()}}{
Read signature cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_sigstsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-read_qcsum"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-read_qcsum}{}}}
\subsection{Method \code{read_qcsum()}}{
Read \code{qc_summary.tsv.gz} cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$read_qcsum(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_sash-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_sash-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_sash$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
