% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sash.R
\name{Wf_sash_download_tidy_write}
\alias{Wf_sash_download_tidy_write}
\title{sash Download Tidy and Write}
\usage{
Wf_sash_download_tidy_write(
  path,
  SubjectID,
  SampleID_tumor,
  SampleID_normal,
  outdir,
  format = "rds",
  max_files = 1000,
  ica_token = Sys.getenv("ICA_ACCESS_TOKEN"),
  regexes = NULL,
  dryrun = FALSE
)
}
\arguments{
\item{path}{Path to directory with raw workflow results (from GDS, S3, or
local filesystem).}

\item{SubjectID}{The SubjectID of the sample.}

\item{SampleID_tumor}{The SampleID of the tumor sample.}

\item{SampleID_normal}{The SampleID of the normal sample.}

\item{outdir}{Path to output directory.}

\item{format}{Format of output files.}

\item{max_files}{Max number of files to list.}

\item{ica_token}{ICA access token (def: $ICA_ACCESS_TOKEN env var).}

\item{regexes}{Tibble with file \code{regex} and \code{fun}ction to parse it.}

\item{dryrun}{If TRUE, just list the files that will be downloaded (don't
download them).}
}
\value{
List where each element is a tidy tibble of a sash file.
}
\description{
Downloads files from the \code{sash} workflow and writes them in a tidy format.
}
\examples{
\dontrun{
SubjectID <- "SBJ03043"
SampleID_tumor <- "PRJ230004"
p1_gds <- glue("gds://production/analysis_data/{SubjectID}/umccrise")
p <- file.path(p1_gds, "20240830ec648f40/L2300064__L2300063")
outdir <- file.path(sub("gds:/", "~/icav1/g", p))
token <- Sys.getenv("ICA_ACCESS_TOKEN")
d <- Wf_sash_download_tidy_write(
  path = p, SubjectID = SubjectID, SampleID_tumor = SampleID_tumor,
  outdir = outdir,
  dryrun = F
)
}
}
