% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umccrise.R
\name{Wf_umccrise}
\alias{Wf_umccrise}
\title{Wf_umccrise R6 Class}
\description{
Reads and writes tidy versions of files from the \code{umccrise} workflow
}
\examples{
\dontrun{
#---- LOCAL ----#
p1_local <- "~/icav1/g/production/analysis_data"
p <- file.path(p1_local, "SBJ01155/umccrise/202408300c218043/L2101566__L2101565")
um1 <- Wf_umccrise$new(path = p)
um1$list_files(max_files = 10)
um1$list_files_filter_relevant(max_files = 100)
}

}
\section{Super class}{
\code{\link[dracarys:Wf]{dracarys::Wf}} -> \code{Wf_umccrise}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf_umccrise-new}{\code{Wf_umccrise$new()}}
\item \href{#method-Wf_umccrise-tidy}{\code{Wf_umccrise$tidy()}}
\item \href{#method-Wf_umccrise-read_conpairmultiqc}{\code{Wf_umccrise$read_conpairmultiqc()}}
\item \href{#method-Wf_umccrise-clone}{\code{Wf_umccrise$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="download_files"><a href='../../dracarys/html/Wf.html#method-Wf-download_files'><code>dracarys::Wf$download_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files"><a href='../../dracarys/html/Wf.html#method-Wf-list_files'><code>dracarys::Wf$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files_filter_relevant"><a href='../../dracarys/html/Wf.html#method-Wf-list_files_filter_relevant'><code>dracarys::Wf$list_files_filter_relevant()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="print"><a href='../../dracarys/html/Wf.html#method-Wf-print'><code>dracarys::Wf$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="tidy_files"><a href='../../dracarys/html/Wf.html#method-Wf-tidy_files'><code>dracarys::Wf$tidy_files()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-new"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-new}{}}}
\subsection{Method \code{new()}}{
Create a new Wf_umccrise object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$new(path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Output directory path with results.}

\item{\code{wname}}{Name of workflow.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-tidy"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-tidy}{}}}
\subsection{Method \code{tidy()}}{
Tidy up the output files from umccrise
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$tidy(indir, out_files)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indir}}{Path to the \code{umccrise} directory.}

\item{\code{out_files}}{Tibble with file path, basename, and size.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_conpairmultiqc"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_conpairmultiqc}{}}}
\subsection{Method \code{read_conpairmultiqc()}}{
Read multiqc_conpair.txt file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_conpairmultiqc(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
