% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umccrise.R
\name{Wf_umccrise}
\alias{Wf_umccrise}
\title{Wf_umccrise R6 Class}
\description{
Reads and writes tidy versions of files from the \code{umccrise} workflow
}
\examples{
\dontrun{

#---- LOCAL ----#
SubjectID <- "SBJ03043"
SampleID_tumor <- "PRJ230004"
prefix <- glue("{SubjectID}__{SampleID_tumor}")
p1_local <- "~/icav1/g/production/analysis_data"
p <- file.path(p1_local, "SBJ03043/umccrise/20240830ec648f40/L2300064__L2300063")
um1 <- Wf_umccrise$new(path = p, SubjectID = SubjectID, SampleID_tumor = SampleID_tumor)
um1$list_files(max_files = 10)
um1$list_files_filter_relevant()
d <- um1$download_files(max_files = 1000, dryrun = F)
d_tidy <- um1$tidy_files(d)
d_write <- um1$write(
  d_tidy,
  outdir = file.path(p, "dracarys_tidy"),
  prefix = glue("{SubjectID}__{SampleID_tumor}"),
  format = "tsv"
)

#---- GDS ----#
SubjectID <- "SBJ03606"
SampleID_tumor <- "PRJ230726"
SampleID_normal <- "PRJ230725"
prefix <- glue("{SubjectID}__{SampleID_tumor}")
p1_gds <- "gds://production/analysis_data"
p <- file.path(p1_gds, "SBJ03606/umccrise/20240829d11e13b0/L2300828__L2300827")
outdir <- file.path(sub("gds:/", "~/icav1/g", p))
token <- Sys.getenv("ICA_ACCESS_TOKEN")
um2 <- Wf_umccrise$new(
  path = p, SubjectID = SubjectID,
  SampleID_tumor = SampleID_tumor, SampleID_normal = SampleID_normal
)
um2$list_files(max_files = 8)
um2$list_files_filter_relevant(ica_token = token, max_files = 500)
d <- um2$download_files(
  outdir = outdir, ica_token = token,
  max_files = 1000, dryrun = F
)
d_tidy <- um2$tidy_files(d)
d_write <- um2$write(
  d_tidy,
  outdir = file.path(outdir, "dracarys_tidy"),
  prefix = glue("{SubjectID}__{SampleID_tumor}"),
  format = "tsv"
)
}

}
\section{Super class}{
\code{\link[dracarys:Wf]{dracarys::Wf}} -> \code{Wf_umccrise}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{SubjectID}}{The SubjectID of the sample.}

\item{\code{SampleID_tumor}}{The SampleID of the tumor sample.}

\item{\code{SampleID_normal}}{The SampleID of the normal sample.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Wf_umccrise-new}{\code{Wf_umccrise$new()}}
\item \href{#method-Wf_umccrise-print}{\code{Wf_umccrise$print()}}
\item \href{#method-Wf_umccrise-read_pcgrJson}{\code{Wf_umccrise$read_pcgrJson()}}
\item \href{#method-Wf_umccrise-read_hrdChord}{\code{Wf_umccrise$read_hrdChord()}}
\item \href{#method-Wf_umccrise-read_hrdHrdetect}{\code{Wf_umccrise$read_hrdHrdetect()}}
\item \href{#method-Wf_umccrise-read_sigsTsv}{\code{Wf_umccrise$read_sigsTsv()}}
\item \href{#method-Wf_umccrise-read_qcSum}{\code{Wf_umccrise$read_qcSum()}}
\item \href{#method-Wf_umccrise-read_conpair}{\code{Wf_umccrise$read_conpair()}}
\item \href{#method-Wf_umccrise-clone}{\code{Wf_umccrise$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="download_files"><a href='../../dracarys/html/Wf.html#method-Wf-download_files'><code>dracarys::Wf$download_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files"><a href='../../dracarys/html/Wf.html#method-Wf-list_files'><code>dracarys::Wf$list_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="list_files_filter_relevant"><a href='../../dracarys/html/Wf.html#method-Wf-list_files_filter_relevant'><code>dracarys::Wf$list_files_filter_relevant()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="tidy_files"><a href='../../dracarys/html/Wf.html#method-Wf-tidy_files'><code>dracarys::Wf$tidy_files()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dracarys" data-topic="Wf" data-id="write"><a href='../../dracarys/html/Wf.html#method-Wf-write'><code>dracarys::Wf$write()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-new"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-new}{}}}
\subsection{Method \code{new()}}{
Create a new Wf_umccrise object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$new(
  path = NULL,
  SubjectID = NULL,
  SampleID_tumor = NULL,
  SampleID_normal = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to directory with raw workflow results (from GDS, S3, or
local filesystem).}

\item{\code{SubjectID}}{The SubjectID of the sample.}

\item{\code{SampleID_tumor}}{The SampleID of the tumor sample.}

\item{\code{SampleID_normal}}{The SampleID of the normal sample.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-print"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-print}{}}}
\subsection{Method \code{print()}}{
Print details about the Workflow.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(ignored).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_pcgrJson"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_pcgrJson}{}}}
\subsection{Method \code{read_pcgrJson()}}{
Read \code{pcgr.json.gz} file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_pcgrJson(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_hrdChord"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_hrdChord}{}}}
\subsection{Method \code{read_hrdChord()}}{
Read \code{chord.tsv.gz} cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_hrdChord(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_hrdHrdetect"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_hrdHrdetect}{}}}
\subsection{Method \code{read_hrdHrdetect()}}{
Read \code{hrdetect.tsv.gz} cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_hrdHrdetect(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_sigsTsv"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_sigsTsv}{}}}
\subsection{Method \code{read_sigsTsv()}}{
Read signature cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_sigsTsv(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_qcSum"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_qcSum}{}}}
\subsection{Method \code{read_qcSum()}}{
Read \code{qc_summary.tsv.gz} cancer report file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_qcSum(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-read_conpair"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-read_conpair}{}}}
\subsection{Method \code{read_conpair()}}{
Read multiqc_conpair.txt file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$read_conpair(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Path to file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Wf_umccrise-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Wf_umccrise-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Wf_umccrise$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
